% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{AddClonesToPlot}
\alias{AddClonesToPlot}
\title{AddClonesToPlot}
\usage{
AddClonesToPlot(
  seuratObj,
  plot,
  fieldName = "CloneNames",
  colorField = NA,
  dotColor = NA,
  pt.size = 1
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{plot}{The plot object, such as the result from FeaturePlot()}
}
\description{
Can be used to highlight a set of cells from a seurat plot, such as overlaying specific clonotypes
}
