% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{AddNewMetaColumn}
\alias{AddNewMetaColumn}
\title{Add a new Seurat object metadata column based on current metadata}
\usage{
AddNewMetaColumn(
  seuratObj,
  varname,
  formulavector,
  defaultname,
  enforceFunctionalFormulae = TRUE
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{varname}{The name of the new metadata column}

\item{formulavector}{vector of formulas, e.g.: c(Tcell_NaiveToEffector > 10 ~ "Effector", Tcell_NaiveToEffector < 5 ~ "Naive")}

\item{defaultname}{The default value applied when none of the formulas in formulavec are TRUE}

\item{enforceFunctionalFormulae}{Boolean determining whether or not to sanity check that all defined formulae in formulavector labeled at least one cell in the Seurat Object.}
}
\value{
Updated Seurat object
}
\description{
This function creates a new Seurat object metadata column based on current formulas applied to current metadata columns
}
