% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSeq.R
\name{AppendCiteSeq}
\alias{AppendCiteSeq}
\title{Read/Append CITE-seq/ADT data to a Seurat Object}
\usage{
AppendCiteSeq(
  seuratObj,
  unfilteredMatrixDir,
  normalizeMethod = "dsb",
  datasetId = NULL,
  assayName = "ADT",
  featureMetadata = NULL,
  adtWhitelist = NULL,
  minRowSum = NULL
)
}
\arguments{
\item{seuratObj}{The seurat object where data will be added.}

\item{unfilteredMatrixDir}{The directory holding raw count data, generally the raw_feature_bc_matrix from the cellranger outs folder}

\item{normalizeMethod}{The method for normalization. If NULL, normalization will be skipped.}

\item{datasetId}{If the seurat object includes cells from multiple datasets, and/or if the barcodes were prefixed with a dataset id, This will be used as a prefix for barcodes, and stored in metadata. Also used as the project name for the Seurat object.}

\item{assayName}{The name of the assay to store ADT data.}

\item{featureMetadata}{If is common for the raw barcode to use non-human friendly names (like TotalSeq-C-XXXX). This is an optional dataframe that can be used to override that, and also provide other feature metadata. This dataframe must have rownames that match the rownames of the saved matrix. If this dataframe has the column 'name', this will be used to replace the original rownames on the ADT assay.}

\item{adtWhitelist}{An optional whitelist of ADT names (matching the raw names from the matrix). If provided, the matrix will be subset to just these features}

\item{minRowSum}{If provided, any ADTs (rows) with rowSum below this value will be dropped.}
}
\description{
Read/Append CITE-seq/ADT data to a Seurat Object
}
