% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{AppendCloneProperties}
\alias{AppendCloneProperties}
\title{Append Clone Properties}
\usage{
AppendCloneProperties(seuratObj, tcrClonesFile)
}
\arguments{
\item{seuratObj}{The Seurat object to append clone properties to}

\item{tcrClonesFile}{The 10x clonotypes file for this Seurat object. Single lanes can use the CellRanger/Vloupe contigs CSV file. Merged lanes need to merge these files and modify them to create unique cellbarcodes and clonotype names, such as the code from Rdiscvr::CreateMergedTcrClonotypeFile().}
}
\description{
A function to append clone properties to a Seurat object
}
