% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{AppendPerCellSaturation}
\alias{AppendPerCellSaturation}
\title{Append Per Cell Saturation}
\usage{
AppendPerCellSaturation(seuratObj, molInfoFile)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{molInfoFile}{The 10x molecule_info.h5 file}
}
\description{
Calculate and Append Per-cell Saturation to a Seurat Object
}
