% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{AppendPerCellSaturationInBulk}
\alias{AppendPerCellSaturationInBulk}
\title{Append Per Cell Saturation in Bulk}
\usage{
AppendPerCellSaturationInBulk(seuratObj, molInfoList)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{molInfoList}{A list mapping dataset/assay to the filepath of the 10x molecule_info.h5 file. The format of the names should be DatasetId-Assayname (i.e. 283729-GEX or 384729-HTO)}
}
\description{
Calculate and Append Per-cell Saturation to a Seurat Object, where the seurat object contains multiple input 10x runs
}
