% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{AvgExpression}
\alias{AvgExpression}
\title{Calculate Avg Expression}
\usage{
AvgExpression(seuratObj, groupField, slot = "counts")
}
\arguments{
\item{seuratObj}{The seurat object}

\item{groupField}{The field on which to group. Must be present in seuratObj@meta.data}

\item{slot}{The slot on which to calculate average expression}
}
\description{
This is a wrapper around Seurat::AverageExpression, which computes the AverageExpression per group (all assays), then flattens this to a single data frame. It also saves library size and total cells per group.
}
