% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{CalculatePercentMito}
\alias{CalculatePercentMito}
\title{Calculate Mitochrondrial Percentage}
\usage{
CalculatePercentMito(
  seuratObj,
  mitoGenesPattern = "^MT-",
  annotateMitoFromReference = TRUE,
  outputColName = "p.mito"
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{mitoGenesPattern}{The expression to use when identifying mitochondrial genes}

\item{annotateMitoFromReference}{If true, a list of mitochondrial genes, taken from (https://www.genedx.com/wp-content/uploads/crm_docs/Mito-Gene-List.pdf) will be used to calculate p.mito}

\item{outputColName}{The name of the output column to hold p.mito}
}
\value{
A Seurat object with p.mito calculated.
}
\description{
This will identify mitochrondrial genes and calculate p.mito for each cell
}
