% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{CalculateTcrDiversity}
\alias{CalculateTcrDiversity}
\title{Calculate TCR diversity}
\usage{
CalculateTcrDiversity(inputData, order1 = 1, order2 = 200)
}
\arguments{
\item{inputData}{The a data frame with the columns: sampleId, v_a_gene, v_b_gene, cdr3_a_aa, cdr3_b_aa, and clone_size}

\item{order1}{The minimum order for calculating the generalized Simpson entropy.}

\item{order2}{The maximum order for calculating the generalized Simpson entropy.}
}
\value{
A data frame with the results
}
\description{
Plot a diversity profile for each library in the data
}
