% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{CalculateTcrRepertoireStats}
\alias{CalculateTcrRepertoireStats}
\title{Calculate TCR Repertoire Stats}
\usage{
CalculateTcrRepertoireStats(
  df,
  groupField = "cDNA_ID",
  minCellsRequiredPerChain = 100,
  minClonesRequiredPerChain = 100
)
}
\arguments{
\item{df}{A dataframe containing metadata for a Seurat object}

\item{groupField}{The sample identifier--this is often cDNA_ID}

\item{minCellsRequiredPerChain}{The minimum number of cells for a sample for diversity to be analyzed}

\item{minClonesRequiredPerChain}{The minimum number of clones in a sample for diversity to be analyzed}
}
\value{
A data frame with the results
}
\description{
Calculate diversity metrics for a Seuratobject metadata dataframe
}
