% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{CalculateTcrRepertioreStatsByPopulation}
\alias{CalculateTcrRepertioreStatsByPopulation}
\title{Calculate TCR Repertoire Stats for all T cells in a sample and for specified subsets}
\usage{
CalculateTcrRepertioreStatsByPopulation(
  df,
  groupfield = "cDNA_ID",
  metacolumn = "RIRA_TNK_v2.cellclass",
  populationvec = c("CD4+ T Cells", "CD8+ T Cells"),
  minCellsRequiredPerChain = 100,
  minClonesRequiredPerChain = 100
)
}
\arguments{
\item{df}{A dataframe containing metadata for a Seurat object}

\item{metacolumn}{The metadata column specifying population subsets. RIRA_TNK_v2.cellclass by default}

\item{populationvec}{A vector that defines the T cell populations to be analyzed. A vector with "CD4+ T Cells", "CD8+ T Cells" by default}

\item{minCellsRequiredPerChain}{The minimum number of cells for a sample for diversity to be analyzed}

\item{minClonesRequiredPerChain}{The minimum number of clones in a sample for diversity to be analyzed}

\item{groupField}{The sample identifier--this is often cDNA_ID}
}
\value{
A data frame with the results
}
\description{
Calculate diversity metrics for a populations in a Seuratobject metadata dataframe
}
