% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSeq.R
\name{CalculateUcellPerFeature}
\alias{CalculateUcellPerFeature}
\title{CalculateUcellPerFeature}
\usage{
CalculateUcellPerFeature(
  seuratObj,
  columnPrefix = NULL,
  assayName = "ADT",
  ncores = 1,
  ignoreCachedRanks = TRUE
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{columnPrefix}{An optional prefix to be applied to the resulting column (which is otherwise the feature name)}

\item{assayName}{The name of the assay}

\item{ncores}{Passed directly to AddModuleScore_UCell}

\item{ignoreCachedRanks}{If true, any previously cached ranked from UCell will be ignored}
}
\value{
A modified Seurat object.
}
\description{
This will iterate all features of the given assay and calculate a per-feature Ucell score
}
