% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdGenes}
\alias{cdGenes}
\title{Table mapping gene symbol to CD gene name}
\format{
A data frame with 394 rows and 4 variables:
\describe{
  \item{GeneSymbol}{Official GeneSymbol}
  \item{ApprovedName}{A description}
  \item{PreviousSymbols}{Previous symbols.  This is the field typically used for aliasing gene symbols}
  \item{Synonyms}{Additional aliases}
  ...
}
}
\source{
\url{https://www.genenames.org/data/genegroup/#!/group/471}
}
\usage{
cdGenes
}
\description{
Table mapping gene symbol to CD gene name
}
\keyword{datasets}
