% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{CellCycleScoring_UCell}
\alias{CellCycleScoring_UCell}
\title{CellCycleScoring_UCell}
\usage{
CellCycleScoring_UCell(
  seuratObj,
  outputFieldName = "Phase",
  min.genes = 10,
  s.genes = GetSPhaseGenes(),
  g2m.genes = GetG2MGenes(TRUE),
  assayName = "RNA",
  min.score.G2M = 0.02,
  min.score.S = 0.02,
  facetField = "ClusterNames_0.2",
  ncores = 1
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{outputFieldName}{The name of the meta.data field to store results}

\item{min.genes}{If less than min.genes are shared between the seurat object and the reference cell cycle genes, this method will abort.}

\item{s.genes}{A vector of features associated with S phase}

\item{g2m.genes}{A vector of features associated with G2M phase}

\item{assayName}{The name of the assay}

\item{min.score.G2M}{The minimum UCell score needed to consider a cell positive for G2M}

\item{min.score.S}{The minimum UCell score needed to consider a cell positive for S}

\item{facetField}{If not NA, a plot will be produced summarizing S.Core and G2M.Score, faceted by this value.}

\item{ncores}{Passed directly to AddModuleScore_UCell}
}
\value{
A modified Seurat object.
}
\description{
Similar to Seurat's CellCycleScoring, except using UCell AddModuleScore_UCell instead of Seurat's AddModuleScore
}
