% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSeq.R
\name{CiteSeqDimRedux.PCA}
\alias{CiteSeqDimRedux.PCA}
\title{Perform DimRedux on CiteSeq using PCA input}
\usage{
CiteSeqDimRedux.PCA(
  seuratObj,
  assayName = "ADT",
  print.plots = TRUE,
  performClrNormalization = TRUE,
  doUMAP = TRUE,
  dimsToUse = NULL,
  adtWhitelist = NULL,
  adtBlacklist = NULL
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{assayName}{The name of the assay holding the ADT data.}

\item{print.plots}{If true, QC plots will be printed}

\item{performClrNormalization}{If true, Seurat's CLR normalization will be performed. Otherwise this expected data to be pre-normalized}

\item{doUMAP}{If true, RunUMAP will be performed}

\item{dimsToUse}{An optional vector of PCs to use in tSNE/UMAP}

\item{adtWhitelist}{An optional vector of ADTs that will be included in PCA. If NULL, all will be used.}

\item{adtBlacklist}{An optional vector of ADTs that will be excluded from PCA. If NULL, all will be used.}
}
\description{
Perform DimRedux on CiteSeq
}
