% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSeq.R
\name{CiteSeqDimRedux}
\alias{CiteSeqDimRedux}
\title{Perform DimRedux on CiteSeq}
\usage{
CiteSeqDimRedux(
  seuratObj,
  assayName = "ADT",
  dist.method = "euclidean",
  print.plots = TRUE,
  performClrNormalization = TRUE,
  doUMAP = TRUE
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{assayName}{The name of the assay holding the ADT data.}

\item{dist.method}{The method, passed to dist()}

\item{print.plots}{If true, QC plots will be printed}

\item{performClrNormalization}{If true, Seurat's CLR normalization will be performed. Otherwise this expected data to be pre-normalized}

\item{doUMAP}{If true, RunUMAP will be performed on the ADT distance matrix.}
}
\description{
Perform DimRedux on CiteSeq
}
