% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{ClrNormalizeByGroup}
\alias{ClrNormalizeByGroup}
\title{ClrNormalizeByGroup}
\usage{
ClrNormalizeByGroup(seuratObj, groupingVar, assayName = "ADT", margin = 2)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{groupingVar}{The variable to use to partition the data}

\item{assayName}{The name of the assay}

\item{margin}{Passed directly to}
}
\description{
This subsets the input object based on a variable (like dataset), and performs CLR normalization per-group, then combines them
}
