% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{CreateStudyWideBarPlot}
\alias{CreateStudyWideBarPlot}
\title{CreateStudyWideBarPlot}
\usage{
CreateStudyWideBarPlot(
  pairwise_de_results,
  pvalue_threshold = 0.05,
  logFC_threshold = 1,
  LikeVsLike = F,
  LikeVsLikeGroupFieldIndex = NULL,
  facetLikeVsLike = T
)
}
\arguments{
\item{pairwise_de_results}{The result of RunPairwiseContrasts().}

\item{pvalue_threshold}{The pvalue/FDR threshold used to filter the differential expression tables for significance.}

\item{logFC_threshold}{Log fold change threshold used to filter the differential expression tables for significance.}

\item{LikeVsLike}{Boolean controlling whether or not one of the contrasts should be held constant in the summarization (i.e., only like cell-types should be considered, since Vaccinated_T cells vs Vaccianted_Myeloid cells will be very different, but not necessarily informative.)}

\item{LikeVsLikeGroupFieldIndex}{The numerical index of the contrast_columns variable supplied to DesignModelMatrix() that should be held constant for like vs like comparisons.}

\item{facetLikeVsLike}{Boolean controlling whether or not the final plot should be faceted in a LikeVsLike contrast. If you have several grouping variables, this could be helpful.}
}
\value{
A ggplot2 bar plot showing the magnitudes of the differentially expressed genes.
}
\description{
Summarize the results of RunPairwiseContrasts() into a series of bar plots, showing the magnitude of differentially expressed genes over many contrasts.
}
