% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR_Seurat.R
\name{DimPlot_SingleRClassLabs}
\alias{DimPlot_SingleRClassLabs}
\title{DimPlot SingleR Class Labels}
\usage{
DimPlot_SingleRClassLabs(
  seuratObject,
  plotIndividually = F,
  datasets = c("hpca")
)
}
\arguments{
\item{seuratObject}{a Seurat object, but if path given, path is prioritized.}

\item{plotIndividually}{If true, two separate plots will be printed.  Otherwise a single plot wil be printed with one above the other}

\item{datasets}{One or more datasets to use as a reference. Allowable values are: hpca, blueprint, dice, monaco, and immgen. See cellDex package for available datasets.}
}
\description{
Create a Dimplot from a Seurat object with SingleR class labels
}
\keyword{Classification}
\keyword{Dimplot}
\keyword{SingleR}
