% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{DoLdaParameterScan}
\alias{DoLdaParameterScan}
\title{DoLdaParameterScan}
\usage{
DoLdaParameterScan(
  seuratObj,
  ntopics = seq(5, 50, by = 5),
  normalizationMethod = "CLR",
  varFeatures = 5000,
  randomSeed = GetSeed(),
  assayName = Seurat::DefaultAssay(seuratObj),
  nCores = 1
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{ntopics}{Passed to runLDA.}

\item{normalizationMethod}{The method used for Seurat::NormalizeData()}

\item{varFeatures}{The number of variable features to use in the LDA model. The more features that are used, the slower the model will run and the more noise that will be introduced, but the model will be more complete in representing your entire dataset.}

\item{randomSeed}{Passed to runLDA seed.number argument}

\item{assayName}{The name of the source assay}

\item{nCores}{The number of cores to use}
}
\description{
This will run LDA on the target assay
}
