% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{.addRegulationInformationAndFilterDEGs}
\alias{.addRegulationInformationAndFilterDEGs}
\title{.addRegulationInformationAndFilterDEGs}
\usage{
.addRegulationInformationAndFilterDEGs(
  tibble,
  logFC_threshold = 1,
  FDR_threshold = 0.05
)
}
\arguments{
\item{tibble}{a (presumably large) tibble from dplyr containing the results of RunFilteredPseudobulkingContrasts}

\item{logFC_threshold}{The minimum value for a gene's absolute value of log fold change to be considered differentially expressed.}

\item{FDR_threshold}{The maximum value for a gene's FDR to be considered differentially expressed.}
}
\value{
a tibble with the uniqueness and direction of regulation for each gene stored in tibble$regulation
}
\description{
a helper function for PseudobulkingBarPlot to tag genes as up- or down-regulated and check for uniqueness of differential gene expression across the contrasts.
}
