% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADTs.R
\name{.CalculateStatistics}
\alias{.CalculateStatistics}
\title{Calculate Statistics for ADT distributions}
\usage{
.CalculateStatistics(
  seuratObj,
  whitelist = NULL,
  adtwhitelist = NULL,
  assay = "ADT",
  layer = "data",
  libraryIdentificationColumn = "cDNA_ID",
  minimumCounts = 200,
  minimumCells = 20,
  plots = T,
  peakdist_sd_ratio = 1,
  peakheight_ratio_threshold = 50
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{whitelist}{A vector of library IDs to run the classification on. If NULL, all libraries will be run.}

\item{adtwhitelist}{An optional vector of ADTs to be triaged and classified.}

\item{assay}{The assay to use.}

\item{layer}{The layer to use.}

\item{libraryIdentificationColumn}{The column in the metadata that identifies the atomic unit of sequencing data (e.g. cDNA library).}

\item{minimumCounts}{A lower bound on the number of reads in the ADT assay to attempt classification.}

\item{minimumCells}{A lower bound on the number of cells recovered from a library to attempt classification.}

\item{plots}{If TRUE, plots will be generated.}

\item{peakdist_sd_ratio}{An upper bound of acceptable peak distance divided by the standard deviation of the ADT.}

\item{peakheight_ratio_threshold}{The threshold for the peak height ratio.}
}
\description{
This function will calculate metrics and produce a dataframe that will be used for classification by .TriageADTs
}
