% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADTs.R
\name{.ClassifyCells}
\alias{.ClassifyCells}
\title{Classify Cells}
\usage{
.ClassifyCells(
  seuratObj,
  df,
  libraryIdentificationColumn,
  assay,
  layer,
  adtwhitelist
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{df}{A dataframe resulting from running .TriageADTs}

\item{libraryIdentificationColumn}{The column in the metadata that identifies the atomic unit of sequencing data (e.g. cDNA library).}

\item{assay}{The assay to use.}

\item{layer}{The layer to use.}

\item{adtwhitelist}{An optional vector of ADTs to be triaged and classified.}
}
\description{
This function will classify cells within a Seurat object based on output from .TriageADTs.
}
