% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA_Alternatives.R
\name{.computeSsgseaScores}
\alias{.computeSsgseaScores}
\title{Compute ssGSEA scores for a gene set}
\usage{
.computeSsgseaScores(gene_set, countMatrix, method)
}
\arguments{
\item{gene_set}{A vector of gene symbols to compute the ssGSEA scores for.}

\item{countMatrix}{A matrix of gene expression values where rows are genes and columns are samples.}

\item{method}{The method to use for ssGSEA. The dts, wass, and ks methods from the twosamples package are supported.}
}
\value{
A vector of ssGSEA scores from a gene set for each sample.
}
\description{
This function computes the ssGSEA scores for a gene set using the twosamples package.
}
