% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{.ConstructDataFrameAndDoStatistics}
\alias{.ConstructDataFrameAndDoStatistics}
\title{.ConstructDataFrameAndDoStatistics}
\usage{
.ConstructDataFrameAndDoStatistics(
  seuratObj,
  colorField = "Population",
  yField = "ClusterNames_0.2",
  groupField = "Timepoint",
  colorValue = "ld-LN-Right",
  colorLabels = c("Left", "Even", "Right"),
  normalizationField = "cDNA_ID",
  sizeFactorField = "SizeFactor",
  maxSizeFactor = 100,
  independentVariableTestField = "Tissue",
  dependentVariableTestField = "SizeFactor"
)
}
\arguments{
\item{seuratObj}{The seurat object that holds the data.}

\item{colorField}{The column of metadata that is used for the colorField}

\item{yField}{The y axis for the dotplots.}

\item{groupField}{The x axis for the dotplots.}

\item{colorValue}{The "value that should be ranked "High" on the color axis.}

\item{colorLabels}{Vector of length 3 that defines the extremes and midpoint of the colorField axis.}

\item{normalizationField}{The metadata column that is used for size factor calculation (equivalently, normalization between different datasets.) Recommended to be cDNA_ID.}

\item{sizeFactorField}{The column name of the seruat object metadata that size factors should be stored in.}

\item{maxSizeFactor}{The maximum allowable SizeFactor before an error is automatically thrown. For instance, a size factor of 100 means you have a group of 8000 cells that you're comparing to 80 cells.}

\item{independentVariableTestField}{This and dependentVariableTestField automatically define a small statistical test to see if your size factors are correlated. Ideally, they should not be if independentVariableTestField is not the same value as normalizationField.}

\item{dependentVariableTestField}{This value should be equal to sizeFactorField initially, but can be changed to interactively see other correlations in the metadata.}
}
\description{
Helper function for makeDotPlot that does the dataframe construction and statistics.
}
