% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{.DoMergeSimple}
\alias{.DoMergeSimple}
\title{doMergeSimple}
\usage{
.DoMergeSimple(seuratObjs, nameList, projectName)
}
\arguments{
\item{seuratObjs}{A list of seurat objects, optionally named (in which case these will be used as dataset names). Also can use SplitObject(, split.by =)}

\item{nameList}{A list of names from MergeSeuratObjs()}

\item{projectName}{The projectName to pass to Seurat}
}
\description{
An internal method to do simple merging of seurat objects from a list of them.
}
