% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{.GenerateSizeFactor}
\alias{.GenerateSizeFactor}
\title{.GenerateSizeFactor}
\usage{
.GenerateSizeFactor(
  seuratObj,
  normalizationField = "cDNA_ID",
  sizeFactorField = "SizeFactor",
  maxSizeFactor = 100
)
}
\arguments{
\item{normalizationField}{The metadata column that is used for size factor calculation (equivalently, normalization between different datasets.) Recommended to be cDNA_ID.}

\item{sizeFactorField}{The column name of the seruat object metadata that size factors should be stored in.}

\item{maxSizeFactor}{The maximum allowable SizeFactor before an error is automatically thrown. For instance, a size factor of 100 means you have a group of 8000 cells that you're comparing to 80 cells.}
}
\description{
An extremely overloaded function that calculates statistics and enrichment in Seurat Objects. Please see an example dot plot before using this function.
}
