% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{.PrepareDataForClassificationGlm}
\alias{.PrepareDataForClassificationGlm}
\title{.PrepareDataForClassificationGlm}
\usage{
.PrepareDataForClassificationGlm(
  seuratObj,
  assay = "RNA",
  layer = "scale.data",
  rescale = TRUE,
  numberOfVariableFeatures = NULL,
  excludeVariableGenes = T,
  metadataVariableForClassification = NULL
)
}
\arguments{
\item{seuratObj}{Input seurat object used for training a glmnet model.}

\item{assay}{Seurat Object's assay}

\item{layer}{layer within the Seurat object assay. Recommended to be "scale.data".}

\item{rescale}{The feature selection will optimize for "heatmap-interpretable genes" so the features are intended to be scaled. If TRUE, this will rescale the variable features.}

\item{numberOfVariableFeatures}{A parameter to select how many features should be selected as variable for scaling, by default, all genes will be used.}

\item{excludeVariableGenes}{A boolean option to exclude variable genes from the analysis. Only the Variable Gene Exclusion list (v2) from RIRA is supported.}

\item{metadataVariableForClassification}{The metadata feature to be classified. If non-binary, then multinomial regression will automatically be performed.}
}
\value{
a matrix with the special characters removed from the gene names.
}
\description{
a helper function for FitClassificationGlm to remove special characters from the gene names.
}
