% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{.RemoveSpecialCharacters}
\alias{.RemoveSpecialCharacters}
\title{RemoveSpecialCharacters}
\usage{
.RemoveSpecialCharacters(vector_of_metadata_values)
}
\arguments{
\item{vector_of_metadata_values}{A vector of metadata values to be sanitized}
}
\value{
A vector of metadata values with special characters removed.
}
\description{
This function is used to remove special characters from metadata values. This ensures that the seuratObj metadata columns are formatted to match entries in the design matrix.
}
