% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{.SeuratToCoNGA}
\alias{.SeuratToCoNGA}
\title{.SeuratToCoNGA}
\usage{
.SeuratToCoNGA(
  seuratObj,
  tcrFileFromRdiscvr,
  seuratToCongaDir,
  assayName = "RNA"
)
}
\arguments{
\item{seuratObj}{The Seurat object to be written into a 10x file format}

\item{tcrFileFromRdiscvr}{The 10x clonotypes file for this run created by Rdiscvr::CreateMergedTcrClonotypeFile().}

\item{seuratToCongaDir}{The local path to write the output files.}

\item{assayName}{The name of the gene expression data assay.}
}
\description{
A wrapper function to prepare a Seurat object for Conga.
}
