% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{.swapContrast}
\alias{.swapContrast}
\title{.swapContrast}
\usage{
.swapContrast(tibble)
}
\arguments{
\item{tibble}{a large tibble from dplyr containing the results of RunFilteredPseudobulkingContrasts called within PseudobulkingDEHeatmap.}
}
\value{
a tibble with the positive and negative contrast directions swapped.
}
\description{
a helper function for PseudobulkingDEHeatmap to swap the "positive" and "negative" sides of the contrasts such that the desired negative contrast value is correctly designated as negative.
}
