% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADTs.R
\name{.TriageADTs}
\alias{.TriageADTs}
\title{Triage ADTs for classification}
\usage{
.TriageADTs(
  df,
  minimumCells,
  peakdist_sd_ratio,
  peakheight_ratio_threshold,
  xdist_ratio_threshold
)
}
\arguments{
\item{df}{A dataframe resulting from running .CalculateStatistics on a Seurat object}

\item{minimumCells}{A lower bound on the number of cells recovered from a library to attempt classification.}

\item{peakdist_sd_ratio}{An upper bound of acceptable peak distance divided by the standard deviation of the ADT.}

\item{peakheight_ratio_threshold}{The threshold for the peak height ratio.}

\item{xdist_ratio_threshold}{The threshold for the peak1_antimode_distance/peak2_antimode_distance ratio.}
}
\description{
This function will take in a dataframe and make calls on bimodality
}
