% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{.UpdateGeneModel}
\alias{.UpdateGeneModel}
\title{Update Gene Model}
\usage{
.UpdateGeneModel(features, predictions = T)
}
\arguments{
\item{features}{a vector of features to be updated to more common gene IDS}

\item{predictions}{Whether or not to include lower quality/speculative updates to the gene model}
}
\description{
Substitutes LOC genes to more common gene IDs
}
