% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{DownsampleSeurat}
\alias{DownsampleSeurat}
\title{Downsample Seurat}
\usage{
DownsampleSeurat(seuratObj, targetCells, subsetFields = NULL, seed = GetSeed())
}
\arguments{
\item{seuratObj}{The seurat object}

\item{targetCells}{The desired cell number to retain per unit of data. If a subsetField is provided, the string 'min' can also be used, in which case}

\item{subsetFields}{If provided, data will be grouped by these fields, and up to targetCells will be retained per group}

\item{seed}{The random seed}
}
\description{
Downsample a seurat object, either globally or subset by a field
}
