% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotyping.R
\name{ExpandGeneList}
\alias{ExpandGeneList}
\title{ExpandGeneList}
\usage{
ExpandGeneList(genes, verbose = TRUE)
}
\arguments{
\item{genes}{A vector of genes or gene set names}

\item{verbose}{Whether to log information about matches}
}
\description{
Takes an input gene list and identifies any entries matching registered gene sets. Those will be expanded to the full gene list.
}
