% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{FeaturePlotAcrossReductions}
\alias{FeaturePlotAcrossReductions}
\title{FeaturePlot Across Reductions}
\usage{
FeaturePlotAcrossReductions(
  seuratObj,
  features,
  reductions = c("tsne", "umap", "wnn.umap"),
  plotsPerRow = 2
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{features}{A vector of features to plot}

\item{reductions}{The list of reductions to plot}
}
\description{
Create FeaturePlots for the selected reductions
}
