% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSeq.R
\name{Filter.ADT.nCountADT}
\alias{Filter.ADT.nCountADT}
\title{Fiter ADTs based on nCountADT}
\usage{
Filter.ADT.nCountADT(seuratObj, min.thresh = 0, max.thresh = NULL)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{min.thresh}{minimum ADT value, default 0 to remove cells with (> thresh) no ADT counts at all}

\item{max.thresh}{maximum nCountADT, default to NULL so to not clip cells (< thresh)}
}
\description{
The sum of all ADT reads per cells i.e., nCountADT is used to fiter
}
