% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{FilterCloneNames}
\alias{FilterCloneNames}
\title{FilterCloneNames}
\usage{
FilterCloneNames(seuratObj, minValue)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{minValue}{Filters clones not present in at least this many cells}
}
\description{
Filter Clone Names
}
