% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{FilterPseudobulkContrasts}
\alias{FilterPseudobulkContrasts}
\title{FilterPseudobulkContrasts}
\usage{
FilterPseudobulkContrasts(
  logic_list = NULL,
  design = NULL,
  use_require_identical_logic = T,
  require_identical_fields = NULL,
  filtered_contrasts_output_file = "./filtered_contrasts.tsv"
)
}
\arguments{
\item{logic_list}{The list that defines the study design. Please see examples for the setup and use. This list be a list of lists where each sub-list has three entries. The first defines the metadata field (e.g. Tissue or Timepoint) to which the logic gate will be applied. The second defines the logic gate that will be used for that field (one of: all (AND), any (OR), xor, nand, nor, or xnor. The third defines the specific value of the metadata field that will be tested for equivalence against the positive and negative contrasts within the gate.}

\item{design}{a design/model matrix returned by DesignModelMatrix().}

\item{use_require_identical_logic}{Whether or not to apply require_identical logic using require_identical_fields. It is possible to use logic gates to keep metadata fields constant, but using this saves time & effort/iteration.}

\item{require_identical_fields}{The metadata columns of the SeuratObj that you wish to keep constant. For example: defining "Tissue" in this vector would filter all contrasts comparing Liver to PBMC, as their tissue is not identical.}

\item{filtered_contrasts_output_file}{The file to write the list of filtered contrasts to (one pair of samples per row, separated into their metadata fields).}
}
\value{
A dataframe of pairwise contrasts, with column names indicating the directionality (positive vs negative) and associated metadata column (inferred by contrast_columns).
}
\description{
This is designed to accept a study design defined by a series of logical gates applied to metadata fields, test them for equivalence to some criterion, and then filter pairwise contrasts accordingly.
}
\examples{
\dontrun{
#Set up design matrix
design <- DesignModelMatrix(pseudobulked_seuratObj, 
                            contrast_columns = c('Timepoint', 'Vaccine', 'Challenge', 'Tissue', 'cell_type', 'SampleType'), 
                            sampleIdCol = 'cDNA_ID')


#Form the study design. This should be a list of gates (each specified as a list itself) with three elements. 
#The first element defines the metadata column or field that will be used for filtering. 
#The second defines the logic gate (AND -> any, OR -> any, xor, xnor, nand, nor). 
#The third defines the specific value of the metadata field that will be tested for equivalence against the positive and negative contrasts within the gate.
logic_list <- list(
                    list('Challenge', 'xor', 'Mock.challenged'), 
                    list('SampleType', 'all', 'Necropsy')
            )


#Finally, enumerate all possible contrasts, then filter and return them according to the study design defined by logic_list.
#If you want to perform "Like-vs-Like" contrasts (i.e. T cell vs T cell within a cell_type metadata field), specify those columns as required to be identical and set use_require_identical_logic to TRUE.
filtered_contrasts <- FilterPseudobulkContrasts(logic_list = logic_list, 
                                                design = design, 
                                                use_require_identical_logic = T, 
                                                require_identical_fields = c('Tissue','cell_type', 'SampleType'), 
                                                filtered_contrasts_output_file = './filtered_contrasts.tsv')
}
}
