% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{FilterRawCounts}
\alias{FilterRawCounts}
\title{Perform Cell Filtering on the cells of the seurat object}
\usage{
FilterRawCounts(
  seuratObj,
  nCount_RNA.high = 20000,
  nFeature.high = 3000,
  pMito.high = 0.15,
  nCount_RNA.low = 0.99,
  nFeature.low = 200,
  pMito.low = -Inf
)
}
\arguments{
\item{nCount_RNA.high}{Cells with nCount_RNA above this value will be filtered}

\item{nFeature.high}{Cells with nFeature above this value will be filtered}

\item{pMito.high}{Cells with percent mito above this value will be filtered}

\item{nCount_RNA.low}{Cells with nCount_RNA below this value will be filtered}

\item{nFeature.low}{Cells with nFeature below this value will be filtered}

\item{pMito.low}{Cells with percent mito  below this value will be filtered}
}
\value{
The modified seurat object
}
\description{
This will perform filtering on the cells of the seurat object
}
