% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{Find_Markers}
\alias{Find_Markers}
\title{Find_Markers}
\usage{
Find_Markers(
  seuratObj,
  identFields,
  outFile = NULL,
  testsToUse = c("wilcox", "MAST", "DESeq2"),
  numGenesToPrint = 20,
  onlyPos = F,
  pValThreshold = 0.001,
  foldChangeThreshold = 0.5,
  minPct = 0.1,
  minDiffPct = -Inf,
  datasetName = NULL,
  assayName = "RNA",
  verbose = FALSE
)
}
\arguments{
\item{seuratObj}{A seurat object}

\item{identFields}{A vector of grouping fields for DE. Often these are the resolution, computed during FindClustersAndDimRedux()}

\item{outFile}{A file where a table of markers will be saved}

\item{testsToUse}{A vector of tests to be used.  Each will be used to run FindAllMarkers() and the results merged. Available are: wilcox, bimod, roc, t, negbinom, poisson, LR, MAST, DESeq2}

\item{numGenesToPrint}{The number of top markers per cluster to print in a table}

\item{onlyPos}{If true, only positive markers will be saved}

\item{pValThreshold}{Only genes with adjusted p-values below this will be reported}

\item{foldChangeThreshold}{Only genes with log2 fold-change above this will be reported}

\item{minPct}{Only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently expressed.}

\item{minDiffPct}{Only test genes that show a minimum difference in the fraction of detection between the two groups.}

\item{datasetName}{An optional label for this dataset. If provided, this will be appended to the resulting table.}

\item{assayName}{The assay to use.}

\item{verbose}{Passed to Seurat::FindMarkers}
}
\value{
A DT::datatable object with the top markers, suitable for printing
}
\description{
Find_Markers
}
