% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doublets.R
\name{FindDoublets}
\alias{FindDoublets}
\title{Find Doublets using scDblFinder}
\usage{
FindDoublets(seuratObj, assay = "RNA")
}
\arguments{
\item{seuratObj}{The seurat object}

\item{assay}{The assay to use}
}
\description{
Find Doublets using scDblFinder
}
