% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{FitRegularizedClassificationGlm}
\alias{FitRegularizedClassificationGlm}
\title{FitRegularizedClassificationGlm}
\usage{
FitRegularizedClassificationGlm(
  seuratObj,
  metadataVariableForClassification = NULL,
  rescale = TRUE,
  numberOfVariableFeatures = 3000,
  assay = "RNA",
  layer = "scale.data",
  devianceCutoff = 0.8,
  split = NULL,
  returnModelAndSplits = F
)
}
\arguments{
\item{seuratObj}{a Seurat object}

\item{metadataVariableForClassification}{The metadata feature to be classified. If non-binary, then multinomial regression will automatically be performed.}

\item{rescale}{The feature selection will optimize for "heatmap-interpretable genes" so the features are intended to be scaled. If TRUE, this will rescale the variable features.}

\item{numberOfVariableFeatures}{A parameter to select how many features should be selected as variable for scaling, by default, all genes will be used.}

\item{assay}{Seurat Object's assay}

\item{layer}{layer within the Seurat object assay. Recommended to be "scale.data".}

\item{devianceCutoff}{Tolerance for model error when deciding how much regularization should be performed. 1 = no tolerance for error, 0 = intercept only, no genes used for prediction.}

\item{split}{the option to provide a previous model's training/testing set. This is necessary if you're performing multiple iterations of model fitting.}

\item{returnModelAndSplits}{A boolean option to return a list containing the fitted model and training/testing splits in addition to the useful features.}
}
\value{
A vector of genes useful for classification and, optionally, the model and training/testing sets.
}
\description{
Treating gene expression like a classification problem, this function trains a penalized model to classify a metadata feature.
}
