% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{GetAssayMetadataSlotName}
\alias{GetAssayMetadataSlotName}
\title{GetAssayMetadataSlotName}
\usage{
GetAssayMetadataSlotName(assayObj)
}
\arguments{
\item{assayObj}{The assay object}
}
\description{
Returns the slotname holding the assay metadata, compatible with seurat 4 and 5
}
