% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{GetChiDF}
\alias{GetChiDF}
\title{Create a dataframe with chi-squared statistics of clones}
\usage{
GetChiDF(seuratObj, field1, field2, plot = FALSE)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{field1}{The name of the first column to compare}

\item{field2}{The name of the second column to compare}

\item{plot}{If true, the function will plot a heatmap using pheatmap}
}
\value{
A dataframe with chi-squared statistics of clones
}
\description{
This function creates a dataframe with chi-squared statistics between two metadata fields
}
