% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{GetG2MGenes}
\alias{GetG2MGenes}
\title{Get G2M Phase Genes}
\usage{
GetG2MGenes(alternateGenes = FALSE)
}
\arguments{
\item{alternateGenes}{If true, the genes will be based on: https://hbctraining.github.io/scRNA-seq_online/lessons/cell_cycle_scoring.html}
}
\value{
The default list of G2M phase genes
}
\description{
Get G2M Phase Genes
}
