% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotyping.R
\name{GetGeneSet}
\alias{GetGeneSet}
\title{GetGeneSet}
\usage{
GetGeneSet(name)
}
\arguments{
\item{name}{The name of the gene set}
}
\description{
Returns a vector with the set of genes registered under the provided name.
}
