% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmPort.R
\name{GetImmPortGenes}
\alias{GetImmPortGenes}
\title{Get ImmPort Genes}
\usage{
GetImmPortGenes()
}
\value{
A character vector of uppercase gene symbols.
}
\description{
Returns all unique gene symbols present in the ImmPort gene list.
}
