% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{GetMsigdbGeneSet}
\alias{GetMsigdbGeneSet}
\title{GetMsigdbGeneSet}
\usage{
GetMsigdbGeneSet(msigdbGeneSets = "H")
}
\arguments{
\item{msigdbGeneSets}{a character vector of gene sets that is either a top level msigdb gene set name (e.g. "H" for hallmark or "C2" for curated gene sets), or a common hierarchical gene set in a large top level (e.g. C5;BP for GO:BP annotations.)}
}
\value{
a named list of gene sets fetched by msigdbr.
}
\description{
a slightly extended escape::getGeneSet wrapper to deal with one-step gene set parsing, which can't parse hierarchical gene sets (C5 + GO:BP) and non-hierarchical gene sets (hallmark, C2 itself, etc) at the same time.
}
