% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{GetXYDataFromPlot}
\alias{GetXYDataFromPlot}
\title{GetXYDataFromPlot}
\usage{
GetXYDataFromPlot(plot, cellNames)
}
\arguments{
\item{plot}{The plot object}

\item{cellNames}{The set of cells to export}
}
\description{
Get XY data from a seurat plot
}
