% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA.R
\name{GO_enrichment}
\alias{GO_enrichment}
\title{Plot SDA Cell Scores}
\usage{
GO_enrichment(
  sdaResults,
  components,
  orgDb = "org.Hs.eg.db",
  geneNumber = 100,
  side = "N"
)
}
\arguments{
\item{sdaResults}{The result list generated by SDATools}

\item{components}{A vector of SDA components to inspect}

\item{orgDb}{The Bioconductor OrgDb (string), which is passed directed to clusterProfiler::enrichGO()}

\item{geneNumber}{The number of genes to inspect}

\item{side}{= "N"}
}
\value{
A list mapping component name to the enrichGO results
}
\description{
Plot SDA Cell Scores
}
