% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{HighlightCellsOnSeuratPlot}
\alias{HighlightCellsOnSeuratPlot}
\title{HighlightCellsOnSeuratPlot}
\usage{
HighlightCellsOnSeuratPlot(
  seuratObj,
  plot,
  cellSelectField = "CloneNames",
  colorLegendLabel = "Clone",
  colorField = NA,
  dotColor = NA,
  pt.size = 1,
  useShape = TRUE,
  horizontalLegend = TRUE
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{plot}{The plot object, such as the result from DimPlot()}

\item{cellSelectField}{The name of the field to select which cells to plot}

\item{colorLegendLabel}{This is passed to labs(color = XXX) to label the legend}

\item{colorField}{The name of the field to assign colors to the cells. This provides the option to use a different value from cellSelectField. If colorField and dotColor are NA, cellSelectField will be used.}

\item{dotColor}{An optional string passed to geom_point(color = XX). This will assign all cells the same color. Ignored if colorField is provided.}

\item{pt.size}{The size, passed to geom_point().}

\item{useShape}{If true, the values of colorField will be used to assign shape. Note: if there are more than 6 unique values this will be ignored.}

\item{horizontalLegend}{If true, theme(legend.box = "horizontal") is added to the plot. This can be useful if the original plot also has a legend (such as cluster names)}
}
\description{
Can be used to highlight a set of cells from a seurat plot, such as overlaying specific clonotypes. Note: this uses ggnewscale::new_scale_color to enable multiple color scales. The original aesthetics are suffixed with '_new' (i.e. color_new). This has also only been tested thoroughly with DimPlots.
}
