% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{InspectSeurat}
\alias{InspectSeurat}
\title{Inspect Seurat}
\usage{
InspectSeurat(seuratObj, slotReportSize = 5e+05, commandReportSize = 5e+05)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{slotReportSize}{The size in bytes, above which a slot's size will be logged.}

\item{commandReportSize}{The size in bytes, above which a command's size will be logged.}
}
\description{
Print various information about the size of a seurat object. Can be useful for debugging.
}
