% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{LogNormalizeUsingAlternateAssay}
\alias{LogNormalizeUsingAlternateAssay}
\title{LogNormalizeUsingAlternateAssay}
\usage{
LogNormalizeUsingAlternateAssay(
  seuratObj,
  assayToNormalize,
  assayForLibrarySize = "RNA",
  scale.factor = 10000,
  maxLibrarySizeRatio = 0.006
)
}
\arguments{
\item{assayToNormalize}{The name of the assay to normalize}

\item{assayForLibrarySize}{The name of the assay from which to derive library sizes. This will be added to the library size of assayToNormalize.}

\item{scale.factor}{A scale factor to be applied in normalization}

\item{maxLibrarySizeRatio}{This normalization relies on the assumption that the library size of the assay being normalized in negligible relative to the assayForLibrarySize. To verify this holds true, the method will error if librarySize(assayToNormalize)/librarySize(assayForLibrarySize) exceeds this value}

\item{seuratObject}{The seurat object}
}
\description{
LogNormalizeUsingAlternateAssay
}
