% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{makeDotPlot}
\alias{makeDotPlot}
\title{makeDotPlot}
\usage{
makeDotPlot(
  seuratObj,
  colorValue = "Lung-Right",
  colorField = "Tissue",
  colorLabels = c("Left", "Even", "Right"),
  yField = "ClusterNames_0.2",
  groupField = "Timepoint",
  normalizationField = "cDNA_ID",
  sizeFactorField = "SizeFactor",
  maxSizeFactor = 100,
  independentVariableTestField = "Tissue",
  dependentVariableTestField = "SizeFactor"
)
}
\arguments{
\item{colorValue}{The "value that should be ranked "High" on the color axis.}

\item{colorField}{The column of metadata that is used for the colorField}

\item{colorLabels}{Vector of length 3 that defines the extremes and midpoint of the colorField axis.}

\item{yField}{The y axis for the dotplots.}

\item{groupField}{The x axis for the dotplots.}

\item{normalizationField}{The metadata column that is used for size factor calculation (equivalently, normalization between different datasets.) Recommended to be cDNA_ID.}

\item{sizeFactorField}{The column name of the seruat object metadata that size factors should be stored in.}

\item{maxSizeFactor}{The maximum allowable SizeFactor before an error is automatically thrown. For instance, a size factor of 100 means you have a group of 8000 cells that you're comparing to 80 cells.}

\item{independentVariableTestField}{This and dependentVariableTestField automatically define a small statistical test to see if your size factors are correlated. Ideally, they should not be if independentVariableTestField is not the same value as normalizationField.}

\item{dependentVariableTestField}{This value should be equal to sizeFactorField initially, but can be changed to interactively see other correlations in the metadata.}
}
\description{
An extremely overloaded function that calculates statistics and enrichment in Seurat Objects. Please see an example dot plot before using this function.
}
