% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{MakeEnrichmentDotPlot}
\alias{MakeEnrichmentDotPlot}
\title{MakeEnrichmentDotPlot}
\usage{
MakeEnrichmentDotPlot(
  seuratObj,
  yField = "ClusterNames_0.2",
  xField = "Timepoint",
  colorField = "Tissue",
  colorLabels = NULL,
  extraGroupingFields = NULL,
  normalizationField = "cDNA_ID",
  sizeFactorField = "SizeFactor",
  maxSizeFactor = 100,
  independentVariableTestField = colorField,
  dependentVariableTestField = sizeFactorField,
  facetingField = NULL
)
}
\arguments{
\item{seuratObj}{The seurat object that holds the data.}

\item{yField}{The y axis for the dotplots.}

\item{xField}{The x axis for the dotplots.}

\item{colorField}{The column of metadata that is used for the colorField. This should be a factor with two levels. The highest level, reported by levels(), will be treated as the 'highest' value.}

\item{colorLabels}{Vector of length 2-3 that defines the extremes and midpoint of the colorField axis. If null, the values of colorField will be used.}

\item{extraGroupingFields}{An optional vector of additional fields to use in grouping}

\item{normalizationField}{The metadata column that is used for size factor calculation (equivalently, normalization between different datasets.) Recommended to be cDNA_ID.}

\item{sizeFactorField}{The column name of the seruat object metadata that size factors should be stored in.}

\item{maxSizeFactor}{The maximum allowable SizeFactor before an error is automatically thrown. For instance, a size factor of 100 means you have a group of 8000 cells that you're comparing to 80 cells.}

\item{independentVariableTestField}{This and dependentVariableTestField automatically define a small statistical test to see if your size factors are correlated. Ideally, they should not be if independentVariableTestField is not the same value as normalizationField.}

\item{dependentVariableTestField}{This value should be equal to sizeFactorField initially, but can be changed to interactively see other correlations in the metadata.}

\item{facetingField}{The column name of the seruat object metadata that the end DotPlot should be facet_wrap'd by.}
}
\description{
An extremely overloaded function that calculates statistics and enrichment in Seurat Objects. Please see an example dot plot before using this function. Note the aggregated data can be obtained from the ggplot object (i.e. P1$data)
}
