% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{MergeSplitLayers}
\alias{MergeSplitLayers}
\title{MergeSplitLayers}
\usage{
MergeSplitLayers(seuratObj)
}
\arguments{
\item{seuratObj}{The seurat object}
}
\value{
The updated seurat object
}
\description{
MergeSplitLayers
}
