% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mitoGenes}
\alias{mitoGenes}
\title{A reference set of mitochondrial genes}
\format{
A data frame with 320 values and one variable
\describe{
  \item{Gene}{The gene symbol}
  ...
}
}
\source{
\url{https://www.genedx.com/wp-content/uploads/crm_docs/Mito-Gene-List.pdf}
}
\usage{
mitoGenes
}
\description{
A reference set of mitochondrial genes
}
\keyword{datasets}
