% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{NanoString_Housekeeping_Normalization}
\alias{NanoString_Housekeeping_Normalization}
\title{NanoString_Housekeeping_Normalization}
\usage{
NanoString_Housekeeping_Normalization(
  seuratObj,
  assay = "RNA",
  targetAssayName = "Housekeeping"
)
}
\arguments{
\item{seuratObj}{The Seurat object that holds the data to be normalized.}

\item{assay}{The assay holding the raw data - the counts slot will be assumed to hold raw data.}

\item{targetAssayName}{The assay name that the normalized data should be stored in.}
}
\value{
Returns a Seurat object containing normalized counts stored in the counts slot of the "Housekeeping" assay.
}
\description{
a normalization method that uses the housekeeping genes to generate size factors for normalization.
}
