% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{NormalizeAndScale}
\alias{NormalizeAndScale}
\title{Run the primary seurat processing steps.}
\usage{
NormalizeAndScale(
  seuratObj,
  variableFeatureSelectionMethod = "vst",
  nVariableFeatures = 2000,
  mean.cutoff = c(0.0125, 3),
  dispersion.cutoff = c(0.5, Inf),
  block.size = 1000,
  variableGenesWhitelist = NULL,
  variableGenesBlacklist = NULL,
  scaleVariableFeaturesOnly = TRUE,
  includeCellCycleGenesInScaleData = TRUE
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{variableFeatureSelectionMethod}{The selection method to be passed to FindVariableFeatures()}

\item{nVariableFeatures}{The number of variable features to find}

\item{mean.cutoff}{Passed directly to FindVariableFeatures}

\item{dispersion.cutoff}{Passed directly to FindVariableFeatures}

\item{block.size}{Passed directly to ScaleData}

\item{variableGenesWhitelist}{An optional vector of genes that will be included in PCA, beyond the default VariableFeatures()}

\item{variableGenesBlacklist}{An optional vector of genes that will be excluded from PCA, beyond the default VariableFeatures()}

\item{scaleVariableFeaturesOnly}{If true, ScaleData will only be performed on VariableFeatures(), which is governed by FindVariableFeatures, variableGenesWhitelist, and variableGenesBlacklist}

\item{includeCellCycleGenesInScaleData}{If true, cell cycle genes will always be included in the features passed to ScaleData().}
}
\value{
A modified Seurat object.
}
\description{
This is the primary entry point for processing scRNAseq data with Seurat
}
