% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetAnalysis.R
\name{PathwayEnrichment}
\alias{PathwayEnrichment}
\title{PathwayEnrichment}
\usage{
PathwayEnrichment(
  seuratObj,
  groupField,
  msigdbSpecies = "Macaca mulatta",
  msigdbCategory = "H",
  scoreType = "pos",
  selectedPathways = c("HALLMARK_TGF_BETA_SIGNALING", "HALLMARK_GLYCOLYSIS",
    "HALLMARK_IL2_STAT5_SIGNALING", "HALLMARK_INFLAMMATORY_RESPONSE",
    "HALLMARK_INTERFERON_ALPHA_RESPONSE", "HALLMARK_G2M_CHECKPOINT",
    "HALLMARK_INTERFERON_GAMMA_RESPONSE", "HALLMARK_DNA_REPAIR", "HALLMARK_APOPTOSIS",
    "HALLMARK_IL6_JAK_STAT3_SIGNALING", "HALLMARK_OXIDATIVE_PHOSPHORYLATION",
    "HALLMARK_UNFOLDED_PROTEIN_RESPONSE", "HALLMARK_TNFA_SIGNALING_VIA_NFKB"),
  msigdbSubcategory = NULL
)
}
\arguments{
\item{seuratObj}{The Seurat object containing the data.}

\item{groupField}{Field on which to group to calculate AUC}

\item{msigdbSpecies}{MSigDBSpecies name to retrieve gene sets from. Run msigdbr_species() to see available species.}

\item{msigdbCategory}{MSigDB collection abbreviation, such as H or C1. Run msigdbr_collections() to see available collections.}

\item{scoreType}{This parameter defines the GSEA score type.Possible options are("std","pos", "neg")}

\item{selectedPathways}{A set of pathways to test.}

\item{msigdbSubcategory}{MSigDB sub-collection abbreviation, such as CGP or BP}
}
\value{
A seurat object with NES values for each pathway appended to each cell
}
\description{
Runs GSEA analysis on a seurat object
}
