% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{PerformDifferentialExpression}
\alias{PerformDifferentialExpression}
\title{PerformDifferentialExpression}
\usage{
PerformDifferentialExpression(
  fit,
  contrast,
  contrast_name,
  logFC_threshold = 1,
  FDR_threshold = 0.05,
  test.use = "QLF"
)
}
\arguments{
\item{fit}{The glm fit object generated by PerformGlmFit}

\item{contrast}{The contrast, passed directly to edgeR::glmQLFTest or edgeR::glmLRT}

\item{contrast_name}{A name for the contrast}

\item{logFC_threshold}{The min logFC}

\item{FDR_threshold}{The min FDR to report}

\item{test.use}{Can be either QLF or LRT. QLF runs edgeR::glmQLFTest, while LRT runs edgeR::glmLRT}
}
\value{
A list with the differential_expression results, volano ggplot object and pvalue_dist ggplot object
}
\description{
This is designed to perform DE using edgeR on a pseudobulked seurat object
}
