% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{PerformEmptyDropletFiltering}
\alias{PerformEmptyDropletFiltering}
\title{PerformEmptyDropletFiltering}
\usage{
PerformEmptyDropletFiltering(
  seuratRawData,
  fdrThreshold = 0.001,
  emptyDropNIters = 10000,
  emptyDropsLower = 100,
  useEmptyDropsCellRanger = FALSE,
  nExpectedCells = 8000
)
}
\arguments{
\item{seuratRawData}{Raw data}

\item{fdrThreshold}{FDR threshold, passed directly to PerformEmptyDrops()}

\item{emptyDropNIters}{Number of iterations, passed directly to PerformEmptyDrops()}

\item{emptyDropsLower}{Passed directly to emptyDrops(). The lower bound on the total UMI count, at or below which all barcodes are assumed to correspond to empty droplets.}

\item{useEmptyDropsCellRanger}{If TRUE, will use DropletUtils::emptyDropsCellRanger instead of emptyDrops}

\item{nExpectedCells}{Only applied if emptyDropsCellRanger is selected. Passed to n.expected.cells argument}
}
\value{
Plot
}
\description{
PerformEmptyDropletFiltering
}
