% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{PerformEmptyDrops}
\alias{PerformEmptyDrops}
\title{PerformEmptyDrops}
\usage{
PerformEmptyDrops(
  seuratRawData,
  emptyDropNIters,
  fdrThreshold = 0.01,
  emptyDropsLower = 100
)
}
\arguments{
\item{seuratRawData}{Raw data}

\item{emptyDropNIters}{Number of iterations, passed directly to PerformEmptyDrops()}

\item{fdrThreshold}{FDR threshold, passed directly to PerformEmptyDrops()}

\item{emptyDropsLower}{Passed directly to emptyDrops(). The lower bound on the total UMI count, at or below which all barcodes are assumed to correspond to empty droplets.}
}
\value{
A modified Seurat object.
}
\description{
PerformEmptyDrops
}
