% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{PerformGlmFit}
\alias{PerformGlmFit}
\title{PerformGlmFit}
\usage{
PerformGlmFit(
  seuratObj,
  design,
  test.use = "QLF",
  assayName = "RNA",
  minCountsPerGene = 1
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{design}{The model.matrix object}

\item{test.use}{Can be either QLF or LRT. QLF runs edgeR::glmQLFTest, while LRT runs edgeR::glmLRT}

\item{assayName}{The name of the assay to use}

\item{minCountsPerGene}{Any genes with fewer than this many counts (across samples) will be dropped.}
}
\value{
An edgeR glm object
}
\description{
Creates an edgeR glm object
}
