% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{PerformIntegration}
\alias{PerformIntegration}
\title{PerformIntegration}
\usage{
PerformIntegration(
  seuratObj,
  splitField = "SubjectId",
  nVariableFeatures = 4000,
  nIntegrationFeatures = 3500,
  k.weight = 20,
  dimsToUse = 20,
  integrationFeaturesInclusionList = NULL,
  integrationFeaturesExclusionList = NULL
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{nVariableFeatures}{The number of variable features used for Seurat::FindVariableFeatures}

\item{nIntegrationFeatures}{The number of features for Seurat::SelectIntegrationFeatures}

\item{k.weight}{Passed to Seurat::IntegrateData()}

\item{dimsToUse}{Passed to Seurat::IntegrateData(), as dims = 1:dimsToUse}

\item{integrationFeaturesInclusionList}{An optional vector of genes that will be included in the integration genes}

\item{integrationFeaturesExclusionList}{An optional vector of genes that will be excluded in the integration genes}
}
\value{
A modified Seurat object.
}
\description{
This perform's Seurat's integration on a dataset
}
