% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA.R
\name{Plot_SDAScoresPerFeature}
\alias{Plot_SDAScoresPerFeature}
\title{Plot SDA Cell Scores By Feature}
\usage{
Plot_SDAScoresPerFeature(
  seuratObj,
  sdaResults,
  metadataFeature,
  direction = "Both"
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{sdaResults}{The result list generated by SDATools}

\item{metadataFeature}{The field to plot (present in the seurat object).}

\item{direction}{Either Pos, Neg, or Both}
}
\description{
Plot SDA Cell Scores By Feature
}
