% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{PlotDiversity}
\alias{PlotDiversity}
\title{Plot Diversity}
\usage{
PlotDiversity(
  conga_clones_file = "./clones_file.txt",
  OutputFile = "./diversity_output.csv",
  order1 = 1,
  order2 = 200
)
}
\arguments{
\item{conga_clones_file}{The output clones file from CoNGA (named clones_file.txt).}

\item{OutputFile}{The output file containing the data used for generating diversity profiles.}

\item{order1}{The minimum order for calculating the generalized Simpson entropy.}

\item{order2}{The maximum order for calculating the generalized Simpson entropy.}
}
\description{
Plot a diversity profile for each library in the data
}
\examples{
\dontrun{
  PlotDiversity(conga_clones_file = "./clones_file.txt",
                OutputFile = "./diversity_output.csv", 
                order1 = 1, order2 = 200)
}
}
