% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotyping.R
\name{PlotImmuneMarkers}
\alias{PlotImmuneMarkers}
\title{PlotImmuneMarkers}
\usage{
PlotImmuneMarkers(seuratObj, reductions = c("tsne", "umap"))
}
\arguments{
\item{seuratObj}{A seurat object}

\item{reductions}{Vector of reduction(s) to use}
}
\description{
Generate a set of Seurat FeaturePlots for common immune cell markers
}
