% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotyping.R
\name{PlotMarkerSeries}
\alias{PlotMarkerSeries}
\title{PlotMarkerSeries}
\usage{
PlotMarkerSeries(seuratObj, reductions, features, title, setSize = 4)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{reductions}{The reduction(s) to use}

\item{features}{The features to plot}

\item{title}{The title for the plot}

\item{setSize}{The number of markers per plot}
}
\description{
Generate a set of Seurat FeaturePlots for the provided markers
}
