% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotyping.R
\name{PlotMarkerSet}
\alias{PlotMarkerSet}
\title{PlotMarkerSeries}
\usage{
PlotMarkerSet(seuratObj, reductions, title, features)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{reductions}{The reduction(s) to use}

\item{title}{The title for the plot}

\item{features}{The features to plot}
}
\description{
Generate a labeled FeaturePlot for the provided markers
}
