% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA.R
\name{PlotSdaCellScores}
\alias{PlotSdaCellScores}
\title{Plot SDA Cell Scores}
\usage{
PlotSdaCellScores(sdaResults, seuratObj, fieldNames, scaleMean0 = TRUE)
}
\arguments{
\item{sdaResults}{The result list generated by SDATools}

\item{seuratObj}{A Seurat object}

\item{fieldNames}{A vector of field names (present in the seurat object) to plot. These must be non-numeric fields.}

\item{scaleMean0}{Boolean default T to scale each compoent to mean 0}
}
\description{
Plot SDA Cell Scores
}
