% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{PlotSeuratVariables}
\alias{PlotSeuratVariables}
\title{Plot Seurat Variables}
\usage{
PlotSeuratVariables(
  seuratObj,
  xvar,
  yvar,
  labelDimplot = FALSE,
  reduction = "umap"
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{xvar}{The first variable to summarize (i.e. ClusterNames_0.2)}

\item{yvar}{The second variable to summarize (i.e. SubjectId)}

\item{labelDimplot}{This value is passed directly to the Seurat::DimPlot label argument}

\item{reduction}{The reduction to use for the Seurat::DimPlot}
}
\description{
Create a panel of plots summarizing two variables within the seurat object
}
