% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR.R
\name{PlotTfData}
\alias{PlotTfData}
\title{Plot TF Data}
\usage{
PlotTfData(
  seuratObj,
  assayName = "tfsulm",
  groupField = Seurat::Idents(seuratObj)
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{assayName}{The source assay}

\item{groupField}{The field on which to group the results}
}
\value{
The pheatmap plot object
}
\description{
This will plot the results of decoupleR as a heatmap
}
