% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctour.R
\name{PredictScTourPseudotime}
\alias{PredictScTourPseudotime}
\title{PredictScTourPseudotime}
\usage{
PredictScTourPseudotime(
  seuratObj,
  modelFile,
  outputBasePath = tempdir(),
  outputReductionName = "sctour",
  assayName = "RNA",
  cleanUpIntermediateFiles = TRUE
)
}
\arguments{
\item{seuratObj}{The Seurat object containing the data.}

\item{modelFile}{A path pointing to a trained scTour model.}

\item{outputBasePath}{The directory where the embeddings and pseudotime will be written. These are deleted unless cleanUpIntermediateFiles=TRUE}

\item{outputReductionName}{The assay name in which to store the results}

\item{assayName}{Assay whose data is to be written out with DropletUtils::write10xCounts. Should be "RNA".}

\item{cleanUpIntermediateFiles}{This boolean controls if GEXOutfile, embeddingOutFile, and exclusionJsonPath should be deleted after model training.}
}
\value{
A Seurat Object with pseudotime from the supplied model written into its metadata.
}
\description{
Predicts pseudotime from a trained sctour model.
}
