% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{PseudobulkSeurat}
\alias{PseudobulkSeurat}
\title{Pseudobulk Seurat}
\usage{
PseudobulkSeurat(
  seuratObj,
  groupFields,
  assays = NULL,
  additionalFieldsToAggregate = NULL,
  metaFieldCollapseCharacter = "|",
  identifyAndPseudobulkProliferativeCluster = F,
  proliferativeClusterGroupingVariable = "ClusterNames_0.2",
  proliferativeMetadataColumn = "Proliferating",
  useConditionalProliferatingBlacklist = T,
  conditionalProliferatingBlacklistField = "RIRA_Immune_v2.cellclass",
  conditionalProliferatingBlacklistValues = "Myeloid"
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{groupFields}{The set of fields on which to group}

\item{assays}{The assays to aggregate}

\item{additionalFieldsToAggregate}{An option list of additional fields (which must be numeric). Per field, the mean per group will be computed and stored in the output object.}

\item{metaFieldCollapseCharacter}{The character to use when concatenating metadata fields together to form the sample key field}

\item{identifyAndPseudobulkProliferativeCluster}{A boolean determining whether or not automatic proliferative cluster detection should be attempted.}

\item{proliferativeClusterGroupingVariable}{The metadata field containing groupings of cells to be specified as proliferating or not.}

\item{proliferativeMetadataColumn}{The metadata column where the "ProliferatingCluster" or "NonProliferatingCluster" will be stored in the seurat object's metadata.}

\item{useConditionalProliferatingBlacklist}{Proliferating cell clusters are more common in lymphocytes, so it may be undesirable to perform this on myeloid cells. This boolean allows you to force cells to belong to a "NonProliferatingCluster" based on another metadata column.}

\item{conditionalProliferatingBlacklistField}{The metadata column of the seurat object's metadata that stores the blacklist values to be corrected to "NonProliferatingCluster" regardless of cell phase.}

\item{conditionalProliferatingBlacklistValues}{The values of conditionalProliferatingBlacklistField that should be set to "NonProliferatingCluster" regardless of cell phase.}
}
\value{
An aggregated Seurat object.
}
\description{
Aggregates raw counts in the seurat object, generating a new seurat object where each same has the sum of the counts, grouped by the desired variables
}
