% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{Q3_Normalization}
\alias{Q3_Normalization}
\title{Q3_Normalization}
\usage{
Q3_Normalization(seuratObj, assay = "RNA", targetAssayName = "Q3")
}
\arguments{
\item{seuratObj}{The Seurat object that holds the data to be normalized.}

\item{assay}{The assay holding the raw data - the counts slot will be assumed to hold raw data.}

\item{targetAssayName}{The assay name that the normalized data should be stored in.}
}
\value{
Returns a Seurat object containing normalized counts stored in the counts slot of the Q3 assay.
}
\description{
a normalization method that uses the 75th percentile to find size factors for data normalization.
}
