% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{ReadAndFilter10xData}
\alias{ReadAndFilter10xData}
\title{Read and Filter 10X files.}
\usage{
ReadAndFilter10xData(
  dataDir,
  datasetId,
  datasetName = NULL,
  emptyDropNIters = 10000,
  emptyDropsFdrThreshold = 0.001,
  storeGeneIds = TRUE,
  emptyDropsLower = 100,
  annotateMitoFromReference = TRUE
)
}
\arguments{
\item{dataDir}{The directory holding raw count data, generally the raw_feature_bc_matrix from the cellranger outs folder}

\item{datasetId}{This will be used as a prefix for barcodes, and stored in metadata. Also used as the project name for the Seurat object.}

\item{datasetName}{An optional print-friendly name that will be stored in metadata}

\item{emptyDropNIters}{The number of iterations to use with PerformEmptyDrops()}

\item{emptyDropsFdrThreshold}{The FDR threshold to call cells in emptyDrops()}

\item{storeGeneIds}{If true, a map to translate geneId and name (by default rownames will use gene name)}

\item{emptyDropsLower}{Passed directly to emptyDrops(). The lower bound on the total UMI count, at or below which all barcodes are assumed to correspond to empty droplets.}

\item{annotateMitoFromReference}{If true, a list of mitochondrial genes, taken from (https://www.genedx.com/wp-content/uploads/crm_docs/Mito-Gene-List.pdf) will be used to calculate p.mito}
}
\value{
A Seurat object.
}
\description{
Reads in 10X files using Read10X and filters abberent cells using PerformEmptyDropletFiltering and returns a Seurat object.
}
