% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{RegressCellCycle}
\alias{RegressCellCycle}
\title{Regress Cell Cycle}
\usage{
RegressCellCycle(seuratObj, scaleVariableFeaturesOnly = T, block.size = 1000)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{scaleVariableFeaturesOnly}{If true, ScaleData will only be performed on genes specified in FindVariableFeatures()}

\item{block.size}{Passed directly to ScaleData}
}
\value{
A modified Seurat object.
}
\description{
Regress Cell Cycle
}
