% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{RemoveCellCycle}
\alias{RemoveCellCycle}
\title{Remove Cell Cycle}
\usage{
RemoveCellCycle(seuratObj, min.genes = 10)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{min.genes}{If less than min.genes are shared between the seurat object and the reference cell cycle genes, this method will abort.}
}
\value{
A modified Seurat object.
}
\description{
Remove Cell Cycle
}
