% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotyping.R
\name{RemoveUnchangedOrZero}
\alias{RemoveUnchangedOrZero}
\title{RemoveUnchangedOrZero}
\usage{
RemoveUnchangedOrZero(seuratObj, reduction, features)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{reduction}{The reduction to use}

\item{features}{The features to evaluate}
}
\value{
The updated set of features
}
\description{
Starting with a set of features, remove any that are unchanged or zeros across all cells
}
