% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{RenameUsingCD}
\alias{RenameUsingCD}
\title{Rename a vector of genes using CD nomenclature}
\usage{
RenameUsingCD(inputGenes)
}
\arguments{
\item{inputGenes}{A vector of genes to be aliased.}
}
\description{
This compares a vector of genes to CD nomenclature, based on https://www.genenames.org/data/genegroup/#!/group/471. Any gene symbol that matches will have the CD name appended to the end (i.e. KLRB1 becomes "KLRB1 (CD161)")
}
