% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{RLE_Plot}
\alias{RLE_Plot}
\title{RLE_Plot}
\usage{
RLE_Plot(
  seuratObj,
  assay = "RNA",
  sampleIdentifier = "SegmentDisplayName",
  colorVariable
)
}
\arguments{
\item{seuratObj}{The Seurat object that holds the data to be normalized.}

\item{assay}{The assay holding the raw data - the counts slot will be assumed to hold raw data.}

\item{sampleIdentifier}{The metadata column that defines individual samples.}

\item{colorVariable}{The metadata column that determines the color of the boxplots of RLE. Optimally, this should be some type of phenotypic description.}
}
\value{
Returns a relative log expression plot
}
\description{
a function to compute relative log expression plots to qualitatively test whether or not a normalization successfully normalized the data.
}
