% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctour.R
\name{Run_sctour}
\alias{Run_sctour}
\title{Run sctour}
\usage{
Run_sctour(
  seuratObj = NULL,
  GEX_outfile = NULL,
  meta_outfile = NULL,
  assayName = "RNA",
  exclusionList = NULL,
  exclusion_JSON_path = NULL,
  ptime_file = NULL
)
}
\arguments{
\item{seuratObj}{The Seurat object containing the data to be run using sctour}

\item{GEX_outfile}{The GEX filename used for output of DropletUtils::write10xCounts (.h5 extension).}

\item{meta_outfile}{The filename used for containing the seuratObj metadata.}

\item{assayName}{Assay whose data is to be written out with DropletUtils::write10xCounts. Should be "RNA"}

\item{exclusionList}{List of genes to be excluded from variable genes in sctour. Usually a RIRA GeneSet.}

\item{exclusion_JSON_path}{Filename for the file containing the gene exclusion list (.json extension)}

\item{ptime_file}{Filename for the pseudotime results to be written to (.csv extension)}
}
\value{
A Seurat object with the pseudotime results appended.
}
\description{
Runs sctour on a seurat object
}
\examples{
\dontrun{
  CreateMergedTcrClonotypeFile(seuratObj,
                               outputFile = "./tcrs.csv")
  sctourSeuratObj <- Run_sctour(seuratObj = seuratObj,
                             GEX_outfile = "./GEX.h5",
                             meta_outfile = ./meta.csv, 
                             assayName = "RNA", 
                             exclusionList = RIRA::GetGeneSet('VariableGenes_Exclusion.2'),
                             exclusion_JSON_path = ./exclusionList.json,
                             ptime_file = NULL)
}
}
