% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSeq.R
\name{RunAdtPca}
\alias{RunAdtPca}
\title{Perform PCA on CiteSeq/ADT Data}
\usage{
RunAdtPca(
  seuratObj,
  assayName = "ADT",
  print.plots = TRUE,
  performClrNormalization = TRUE
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{assayName}{The name of the assay holding the ADT data.}

\item{print.plots}{If true, QC plots will be printed}

\item{performClrNormalization}{If true, Seurat's CLR normalization will be performed. Otherwise this expected data to be pre-normalized}
}
\description{
Perform PCA on CiteSeq/ADT Data
}
