% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellBender.R
\name{RunCellBender}
\alias{RunCellBender}
\title{Run CellBender}
\usage{
RunCellBender(
  rawFeatureMatrix,
  expectedCells = 5000,
  totalDropletsIncluded = 20000,
  fpr = 0.01,
  epochs = 150
)
}
\arguments{
\item{rawFeatureMatrix}{The path to 10x's raw_feature_bc_matrix.h5}

\item{expectedCells}{Passed to CellBender --expected-cells}

\item{totalDropletsIncluded}{Passed to CellBender --total-droplets-included}

\item{fpr}{Passed to CellBender --fpr}

\item{epochs}{Passed to CellBender --epochs}
}
\description{
Runs CellBender using the 10x h5 file as input
}
