% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{RunCoNGA}
\alias{RunCoNGA}
\title{Run CoNGA}
\usage{
RunCoNGA(
  features_file,
  tcr_datafile,
  gex_datafile,
  organism,
  outfile_prefix,
  gex_datatype,
  clones_file,
  outfile_prefix_for_qc_plots
)
}
\arguments{
\item{features_file}{textfile with list of most variable genes}

\item{tcr_datafile}{Merged TCR clonotype file}

\item{gex_datafile}{h5 file}

\item{organism}{'human' or 'rhesus'}

\item{outfile_prefix}{prefix to output files}

\item{gex_datatype}{"10x_h5"}

\item{clones_file}{clones file generated by CoNGA}

\item{outfile_prefix_for_qc_plots}{prefix to qc output files}
}
\description{
Runs CoNGA on data generated by Rdiscvr::CreateMergedTcrClonotypeFile
}
