% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{runCoNGA_on_seuratdata_and_subsets}
\alias{runCoNGA_on_seuratdata_and_subsets}
\title{Run CoNGA on Seurat data and subsets}
\usage{
runCoNGA_on_seuratdata_and_subsets(
  seuratObj,
  splitFields,
  outputdir,
  organism,
  gex_datatype
)
}
\arguments{
\item{seuratObj}{input Seurat object}

\item{splitFields}{a vector of metadata columns used for splitting the Seurat object}

\item{outputdir}{file path of directory for output files}

\item{organism}{'human' or 'rhesus'}

\item{gex_datatype}{"10x_h5"}
}
\description{
Runs CoNGA on combo data and on combo data split apart into individual samples
}
