% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSCore.R
\name{RunCsCore}
\alias{RunCsCore}
\title{Run CS-CORE}
\usage{
RunCsCore(
  seuratObj,
  genes_selected = Seurat::VariableFeatures(seuratObj),
  saveFile = NULL
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{genes_selected}{Passed directly to CSCORE::CSCORE()}

\item{saveFile}{An optional file to save an RDS of the CS-CORE results}
}
\value{
The results after WGCNA
}
\description{
This will run CS-CORE on the default assay
}
