% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleR.R
\name{RunDecoupleR}
\alias{RunDecoupleR}
\title{Run decoupleR}
\usage{
RunDecoupleR(
  seuratObj,
  sourceAssay = "RNA",
  targetAssay = "tfsulm",
  organism = "human"
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{sourceAssay}{The source assay}

\item{targetAssay}{The target assay to store results}

\item{organism}{Passed directly to decoupleR::get_collectri()}
}
\value{
The seurat object
}
\description{
This will run decoupleR to infer TF
}
