% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{RunFilteredContrasts}
\alias{RunFilteredContrasts}
\title{RunFilteredContrasts}
\usage{
RunFilteredContrasts(
  seuratObj,
  filteredContrastsFile = NULL,
  filteredContrastsDataframe = NULL,
  design,
  test.use,
  logFC_threshold = 1,
  minCountsPerGene = 1,
  assayName = "RNA"
)
}
\arguments{
\item{filteredContrastsFile}{A file containing the output of FilterPseudobulkContrasts. This can be omitted if you use the dataframe returned by FilterPseudobulkContrasts insetad.}

\item{filteredContrastsDataframe}{A dataframe containing the output of FilterPseudobulkContrasts. This can be omitted if you read the results of FilterPseudobulkContrasts from the file written by the function instead.}

\item{design}{a design/model matrix returned by DesignModelMatrix().}

\item{test.use}{A passthrough argument specifying the statistical test to be run by PerformGlmFit and PerformDifferentialExpression.}

\item{logFC_threshold}{A passthrough argument specifying the log fold change threshold to be used by PerformDifferentialExpression.}

\item{minCountsPerGene}{A passthrough argument specifying the minimum counts a gene must have before it is filtered.}

\item{assayName}{A passthrough argument specifying in which assay the counts are held.}
}
\value{
A list of dataframes containing the differentially expressed genes in each contrast supplied by one of the filteredGenes arguments.
}
\description{
This is the companion function to FilterPseudobulkContrasts. FilterPseudobulkContrasts determines the valuable contrasts relative to your metadata variable of interest, and this function performs the model fitting and collecting the differential expression results.
}
