% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdWGCNA.R
\name{RunHdWGCNA}
\alias{RunHdWGCNA}
\title{Run hdWGCNA}
\usage{
RunHdWGCNA(
  seuratObj,
  wgcna_name,
  groupName,
  groupBy,
  sampleGroupingVariables = c("SubjectId", "CellType"),
  geneSelectionApproach = "fraction",
  fraction = 0.05,
  reductionName = "harmony",
  assayName = "RNA"
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{wgcna_name}{The name of the hdWGCNA experiment}

\item{groupName}{The name of the group of interest in the groupBy column}

\item{groupBy}{The metadata column containing the cell type info.}

\item{sampleGroupingVariables}{Metadata variables on which to group cells for MetacellsByGroups()}

\item{geneSelectionApproach}{The gene selection approach}

\item{fraction}{Fraction of cells that a gene needs to be expressed in order to be included}

\item{reductionName}{Select the dimensionality reduction to perform KNN on}
}
\value{
The modified seurat object
}
\description{
Runs hdWGCNA on a seurat object
}
