% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{RunPairwiseContrasts}
\alias{RunPairwiseContrasts}
\title{RunPairwiseContrasts}
\usage{
RunPairwiseContrasts(fit, test.use, logFC_threshold = 1)
}
\arguments{
\item{fit}{The glm fit object generated by PerformGlmFit}

\item{test.use}{Can be either QLF or LRT. QLF runs edgeR::glmQLFTest, while LRT runs edgeR::glmLRT}

\item{logFC_threshold}{Log fold change threshold pass-through variable to be sent to PerformDifferentialExpression.}
}
\value{
A list with the differential_expression results, volano ggplot object and pvalue_dist ggplot object
}
\description{
A convenience function to iterate and perform pairwise contrasts using the glm object
}
