% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{RunPcaSteps}
\alias{RunPcaSteps}
\title{Run Seurat PCA}
\usage{
RunPcaSteps(seuratObj, npcs = 50, variableGeneTable = NULL)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{npcs}{Number of PCs to use for RunPCA()}

\item{variableGeneTable}{If provided, a table of variable genes will be written to this file}
}
\value{
A modified Seurat object.
}
\description{
Run Seurat PCA
}
