% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMetabolism.R
\name{RunScMetabolism}
\alias{RunScMetabolism}
\title{Run scMetabolism}
\usage{
RunScMetabolism(
  seuratObj,
  method = "AUCell",
  doImputation = FALSE,
  metabolismType = "KEGG",
  plotVariables = c("ClusterNames_0.2", "ClusterNames_0.4", "ClusterNames_0.6"),
  dropPathwaysFromAssays = TRUE
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{method}{Passed to scMetabolism::sc.metabolism.Seurat}

\item{doImputation}{Allows users to choose whether impute their data before metabolism scoring.}

\item{metabolismType}{Either KEGG (85 pathways) or REACTOME (82 pathways)}

\item{plotVariables}{A list of variables used for grouping when generating a DotPlot}

\item{dropPathwaysFromAssays}{If true}
}
\value{
The modified seurat object
}
\description{
Runs scMetabolism on a seurat object
}
