% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA.R
\name{RunSDA}
\alias{RunSDA}
\title{RunSDA}
\usage{
RunSDA(
  seuratObj,
  outputFolder,
  numComps = 50,
  assayName = "RNA",
  randomSeed = GetSeed(),
  minAsinhThreshold = 0.5,
  featureInclusionList = NULL,
  featureExclusionList = NULL,
  minLibrarySize = 50,
  path.sda = "sda_static_linux",
  max_iter = 10000,
  nThreads = 1
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{outputFolder}{The path to save results. There will be subfolders for ./rawData and ./results}

\item{numComps}{Passed to SDAtools::run_SDA(). 30 is a good minimum but depends on input data complexity.}

\item{assayName}{The name of the assay}

\item{randomSeed}{Passed to SDAtools::run_SDA() set_seed}

\item{minAsinhThreshold}{Only features with asinh() on total counts above this value will be included.}

\item{featureInclusionList}{An optional vector of genes that will be included in SDA}

\item{featureExclusionList}{An optional vector of genes that will be excluded from SDA}

\item{minLibrarySize}{Passed to dropsim::normaliseDGE() min_library_size}

\item{path.sda}{The full path to the SDA binary. By default it assumes sda_static_linux in in your $PATH}

\item{max_iter}{Passed directly to SDAtools::run_SDA()}

\item{nThreads}{Passed to SDAtools::run_SDA() num_openmp_threads}
}
\description{
This will run SDA on the target assay
}
