% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSeq.R
\name{RunSeuratWnn}
\alias{RunSeuratWnn}
\title{Perform Seurat WNN on CiteSeq}
\usage{
RunSeuratWnn(seuratObj, dims.list = list(1:30, 1:18))
}
\arguments{
\item{seuratObj}{The seurat object}

\item{dims.list}{Passed directly to Seurat::FindMultiModalNeighbors}
}
\description{
Perform Seurat WNN on CiteSeq
}
