% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleR_Seurat.R
\name{RunSingleR}
\alias{RunSingleR}
\title{Run SingleR For A Seurat Object}
\usage{
RunSingleR(
  seuratObj = NULL,
  datasets = c("hpca", "blueprint", "dice", "monaco"),
  assay = NULL,
  resultTableFile = NULL,
  rawDataFile = NULL,
  minFraction = 0.01,
  showHeatmap = TRUE,
  maxCellsForHeatmap = 20000,
  nThreads = NULL
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{datasets}{One or more datasets to use as a reference. Allowable values are: hpca, blueprint, dice, monaco, and immgen. See cellDex package for available datasets.}

\item{assay}{The assay in the seurat object to use}

\item{resultTableFile}{If provided, a table of results will be saved here}

\item{rawDataFile}{If provided, the complete SingleR results will be saved to this file}

\item{minFraction}{If provided, any labels present with fraction of this or fewer across cells will be converted to Unknown}

\item{showHeatmap}{If true, heatmaps will be generated showing the SingleR calls}

\item{maxCellsForHeatmap}{The heatmap will only be plotted if the total cells is below this number}

\item{nThreads}{If provided, this integer value is passed to SingleR's BPPARAM argument. On windows ths is passed to BiocParallel::SnowParam(). On other OS it is passed to BiocParallel::MulticoreParam()}
}
\value{
The modified seurat object
}
\description{
Compute SingleR classification on a Seurat object
}
