% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tricycle.R
\name{RunTricycle}
\alias{RunTricycle}
\title{Run tricycle}
\usage{
RunTricycle(seuratObj, assay = NULL, species = "human", gname.type = "SYMBOL")
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{assay}{The name of the assay to use. Defaults to Seurat::DefaultAssay()}

\item{species}{The species}

\item{gname.type}{Passed directly to tricycle functions}
}
\value{
The seurat object with results stored in metadata variables
}
\description{
This will run tricycle to calculate cell cycle phase
}
