% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA.R
\name{RunUmapOnSDA}
\alias{RunUmapOnSDA}
\title{Run UMAP using SDA results}
\usage{
RunUmapOnSDA(
  seuratObj,
  dimsToUse = NULL,
  minDimsToUse = NULL,
  reduction = "sda"
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{dimsToUse}{The number of dims to use.  If null, this will be inferred using FindSeuratElbow()}

\item{minDimsToUse}{The minimum numbers of dims to use.  If dimsToUse is provided, this will override.}

\item{reduction}{The name of the source reduction}
}
\value{
A modified Seurat object.
}
\description{
Run UMAP using SDA results
}
