% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vision.R
\name{RunVisionForMSigDB}
\alias{RunVisionForMSigDB}
\title{Run VISION using MSigDB Gene Sets}
\usage{
RunVisionForMSigDB(
  seuratObj,
  metadataCols = c("nCount_RNA", "ClusterNames_0.2", "ClusterNames_0.4",
    "ClusterNames_0.6", "ClusterNames_0.8")
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{metadataCols}{A list of column names to include in VISION analysis.}
}
\description{
Run VISION using MSigDB Gene Sets. Returns an object that can be inspected using: VISION::viewResults(vision.obj)
}
