% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{RUVg_Housekeeping_Normalization}
\alias{RUVg_Housekeeping_Normalization}
\title{RUVg_Housekeeping_Normalization}
\usage{
RUVg_Housekeeping_Normalization(
  seuratObj,
  assay = "RNA",
  targetAssayName = "RUVg",
  k = 1
)
}
\arguments{
\item{seuratObj}{The Seurat object that holds the data to be normalized.}

\item{assay}{The assay holding the raw data - the counts slot will be assumed to hold raw data.}

\item{targetAssayName}{The assay name that the normalized data should be stored in.}

\item{k}{the number of factors of unwanted variation. See section "Choice of number of factors of unwanted variation." in paper cited in the description.}
}
\value{
Returns a Seurat object containing normalized counts stored in the counts slot of the RUVg assay.
}
\description{
a normalization method that uses residuals from a fitted model to Remove Unwanted Variation by leveraging housekeeping genes (https://www.nature.com/articles/nbt.2931).
}
