% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{ScaleFeaturesIfNeeded}
\alias{ScaleFeaturesIfNeeded}
\title{Scale Features If Needed}
\usage{
ScaleFeaturesIfNeeded(seuratObj, toScale, assayName = "RNA")
}
\arguments{
\item{seuratObj}{The seurat object}

\item{toScale}{The list of features to scale.}

\item{assayName}{The name of the assay to use.}
}
\description{
This accepts a list of features and appends them to the target assay, if they are not already scaled.
}
