% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{ScoreCellCycle}
\alias{ScoreCellCycle}
\title{Score Cell Cycle}
\usage{
ScoreCellCycle(seuratObj, min.genes = 10)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{min.genes}{If less than min.genes are shared between the seurat object and the reference cell cycle genes, this method will abort.}
}
\value{
A modified Seurat object.
}
\description{
Score Cell Cycle
}
