% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA.R
\name{SDA_GO_Enrichment}
\alias{SDA_GO_Enrichment}
\title{Compute and score GO enrichment for SDA components}
\usage{
SDA_GO_Enrichment(
  sdaResults,
  components,
  orgDb = "org.Hs.eg.db",
  geneNumber = 100,
  direction = "Both"
)
}
\arguments{
\item{sdaResults}{The result list generated by SDATools}

\item{components}{A vector of SDA components to inspect}

\item{orgDb}{The Bioconductor OrgDb (string), which is passed directed to clusterProfiler::enrichGO()}

\item{geneNumber}{The number of genes to inspect}

\item{direction}{Whether to include positive genes, negative, or both}
}
\value{
A list mapping component name + direction to the enrichGO results
}
\description{
Compute and score GO enrichment for SDA components
}
