% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA.R
\name{SDAToSeuratMetadata}
\alias{SDAToSeuratMetadata}
\title{SDAToSeuratMetadata}
\usage{
SDAToSeuratMetadata(seuratObj, results, plotComponents = TRUE)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{results}{The SDA results list}

\item{plotComponents}{If true, FeaturePlots will be generated for the components}
}
\description{
This will store the SDA scores in the seurat object metadata table
}
