% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA.R
\name{SDAToSeuratReduction}
\alias{SDAToSeuratReduction}
\title{SDAToSeuratReduction}
\usage{
SDAToSeuratReduction(
  seuratObj,
  sdaResults,
  assayName = "RNA",
  reduction.name = "sda",
  reduction.key = "SDA_"
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{sdaResults}{The SDA results list}

\item{assayName}{The source assay}

\item{reduction.name}{The name used for this reduction}

\item{reduction.key}{The key used for this reduction}
}
\description{
This will store the SDA results as a reduction in the seurat object
}
