% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{SpatialNormalizeByGroup}
\alias{SpatialNormalizeByGroup}
\title{SpatialNormalizeByGroup}
\usage{
SpatialNormalizeByGroup(
  seuratObj,
  assay = "RNA",
  normalizationMethod = NULL,
  inferAssayName = T,
  targetAssayName = NULL,
  groupField = "SlideName",
  k = NA
)
}
\arguments{
\item{seuratObj}{The Seurat object that holds the data to be normalized.}

\item{assay}{The assay holding the raw data - the counts slot will be assumed to hold raw data.}

\item{normalizationMethod}{The normalization method to be applied (One of: Q3, Housekeeping, or RUVg) across the batches/groups in groupField.}

\item{targetAssayName}{The assay name that the normalized data should be stored in.}

\item{groupField}{the metadata column that delineates how the samples within Seurat object should be grouped.}

\item{k}{a passthrough variable for the RUVg normalization. Please see RUVg_Housekeeping_Normalization for details about this parameter.}
}
\value{
Returns a Seurat object containing normalized counts stored in the counts slot of the "Housekeeping" assay.
}
\description{
a wrapper function to iteratively apply normalization methods that may be sensitive to batch variation.
}
