% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{SplitSeurat}
\alias{SplitSeurat}
\title{Split Seurat}
\usage{
SplitSeurat(seuratObj, splitField)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{splitField}{The name of the field on which to split the object}

\item{seed}{The random seed}
}
\description{
Split a seurat object, dividing into new objects based on the value of a field
}
