% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{SubsetSeurat}
\alias{SubsetSeurat}
\title{Subset Seurat}
\usage{
SubsetSeurat(seuratObj, expressions)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{expressions}{A vector of expressions to use in selection}
}
\description{
Subset a seurat object, selecting cells that match each expression
}
