% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmPort.R
\name{SubsetSeuratToImmPortGenes}
\alias{SubsetSeuratToImmPortGenes}
\title{Subset Seurat To ImmPort Genes}
\usage{
SubsetSeuratToImmPortGenes(seuratObj, assayName = NULL)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{assayName}{The assay to subset. Defaults to the active default assay.}
}
\value{
The Seurat object with features reduced to the ImmPort gene intersection.
}
\description{
Subsets the features of a Seurat assay to only those present in the ImmPort gene list.
}
