% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctour.R
\name{TrainSctourModel}
\alias{TrainSctourModel}
\title{TrainSctourModel}
\usage{
TrainSctourModel(
  seuratObj,
  modelFileName,
  featureExclusionList = "VariableGenes_Exclusion.2",
  outputBasePath = "./",
  assayName = "RNA",
  cleanUpIntermediateFiles = TRUE,
  outputReductionName = "sctour",
  metadataColName = "pseudotime"
)
}
\arguments{
\item{seuratObj}{The Seurat object containing the data.}

\item{modelFileName}{The model's file name. The full name of the model will be modelFileName.pth}

\item{featureExclusionList}{A vector of gene names to be excluded from variable feature selection in model training. This supports RIRA's ExpandGeneList}

\item{outputBasePath}{A directory that will store the resulting pytorch model after model training.}

\item{assayName}{Assay whose data is to be written out with DropletUtils::write10xCounts. Should be "RNA".}

\item{cleanUpIntermediateFiles}{This boolean controls if GEXOutfile, embeddingOutFile, and exclusionJsonPath should be deleted after model training.}

\item{outputReductionName}{The assay name in which to store the results}

\item{metadataColName}{The name of the meta.data column to store the resulting pseudotime}
}
\value{
A seurat object with pseudotime and dimensional reductions computed by scTour
}
\description{
Trains an sctour model on a Seurat object
}
