% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{UpdateMacaqueMmul10NcbiGeneSymbols}
\alias{UpdateMacaqueMmul10NcbiGeneSymbols}
\title{Update Seurat Object Gene Model From Monkey LOCs to Human Gene IDs}
\usage{
UpdateMacaqueMmul10NcbiGeneSymbols(seuratObj, verbose = T)
}
\arguments{
\item{seuratObj}{The Seurat Object to be updated}

\item{verbose}{This prints the slot and assay number, primarily for debugging once seurat updates.}
}
\description{
Uses UpdateGeneModel to change LOC genes to more common human gene IDs in the RNA assay
}
