% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{WriteSummaryMetrics}
\alias{WriteSummaryMetrics}
\title{WriteSummaryMetrics}
\usage{
WriteSummaryMetrics(seuratObj, file)
}
\arguments{
\item{seuratObj, }{A Seurat object.}

\item{file}{The file where metrics will be written}
}
\description{
WriteSummaryMetrics
}
