% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{CalculateTcrDiversityFromSeurat}
\alias{CalculateTcrDiversityFromSeurat}
\title{Calculate TCR diversity from a seurat object}
\usage{
CalculateTcrDiversityFromSeurat(
  seuratObj,
  groupField,
  order1 = 1,
  order2 = 200
)
}
\arguments{
\item{seuratObj}{A seurat object with the following columns: TRA, TRA_V, TRA_J, TRB, TRB_V, TRB_J}

\item{groupField}{The field that defines sample groups.}

\item{order1}{The minimum order for calculating the generalized Simpson entropy.}

\item{order2}{The maximum order for calculating the generalized Simpson entropy.}
}
\value{
A data frame with the results
}
\description{
Calculate TCR a diversity profile for each library in the data
}
