% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{CellCycleScoring_UCell}
\alias{CellCycleScoring_UCell}
\title{CellCycleScoring_UCell}
\usage{
CellCycleScoring_UCell(
  seuratObj,
  s.features,
  g2m.features,
  set.ident = FALSE,
  assayName = "RNA",
  facetField = "ClusterNames_0.2",
  ncores = 1
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{s.features}{A vector of features associated with S phase}

\item{g2m.features}{A vector of features associated with G2M phase}

\item{set.ident}{If true, sets identity to phase assignments}

\item{assayName}{The name of the assay}

\item{facetField}{If not NA, a plot will be produced summarizing S.Core and G2M.Score, faceted by this value.}

\item{ncores}{Passed directly to AddModuleScore_UCell}
}
\value{
A modified Seurat object.
}
\description{
Similar to Seurat's CellCycleScoring, except using UCell AddModuleScore_UCell instead of Seurat's AddModuleScore
}
