% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{ClusteredDotPlot}
\alias{ClusteredDotPlot}
\title{ClusteredDotPlot}
\usage{
ClusteredDotPlot(
  seuratObj,
  features,
  groupFields = "ClusterNames_0.2",
  assay = "RNA",
  ggplotify = TRUE,
  scaling = "row",
  layer = "data"
)
}
\arguments{
\item{seuratObj}{The Seurat object that holds the data.}

\item{features}{The features to plot.}

\item{groupFields}{The metadata column that is used for grouping.}

\item{assay}{The assay to plot.}

\item{ggplotify}{A boolean that determines if the ComplexHeatmap object should be converted to a ggplot object.}

\item{scaling}{The scaling method for the heatmap. Options are "row", "column", or none.}
}
\description{
A function that generates a clustered dot plot with a heatmap of scaled expression.
}
