% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{ClusteredDotPlot}
\alias{ClusteredDotPlot}
\title{ClusteredDotPlot}
\usage{
ClusteredDotPlot(
  seuratObj,
  features,
  groupFields = "ClusterNames_0.2",
  assay = "RNA",
  scaling = "column",
  layer = "data",
  forceRescaling = FALSE,
  inferDefaultArguments = TRUE,
  printInferredArguments = FALSE,
  ...
)
}
\arguments{
\item{seuratObj}{The Seurat object that holds the data.}

\item{features}{The features to plot.}

\item{groupFields}{The metadata column that is used for grouping.}

\item{assay}{The assay to plot.}

\item{scaling}{The scaling method for the heatmap. Options are "row", "column", or none.}

\item{layer}{The layer of the Seurat object that holds the relevant expression data.}

\item{forceRescaling}{A boolean that determines if the Seurat object should be rescaled to include entries in the features vector if any are missing from the scale.data layer. This might be costly to perform locally.}

\item{inferDefaultArguments}{If TRUE, the function will infer the default arguments for the ComplexHeatmap::Heatmap function.}

\item{printInferredArguments}{Boolean to control optional printing of the arguments inferred by inferDefaultArguments.}

\item{...}{Additional arguments to pass to ComplexHeatmap::Heatmap}
}
\description{
A function that generates a clustered dot plot with a heatmap of scaled expression.
}
\examples{
\dontrun{
#set the seurat Idents for FindAllMarkers
Seurat::Idents(seuratObj) <- "ClusterNames_0.2"

markers <- Seurat::FindAllMarkers(seuratObj)

#filter markers to display the largest cluster identity markers according to average log fold change and differences in pct expression.
strong_markers <- markers[abs(markers$avg_log2FC) > 3 & abs(markers$pct.1 - markers$pct.2) > 0.25, "gene"]

dotPlot <- ClusteredDotPlot(seuratObj, features = strong_markers, groupFields = "ClusterNames_0.2", scaling = 'column')
print(dotPlot)

}
}
