% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{CreateSeuratObj}
\alias{CreateSeuratObj}
\title{Wrapper around Seurat::CreateSeuratObject}
\usage{
CreateSeuratObj(
  seuratData,
  datasetId,
  datasetName = NULL,
  minFeatures = 25,
  minCells = 0,
  mitoGenesPattern = "^MT-",
  annotateMitoFromReference = TRUE
)
}
\arguments{
\item{seuratData}{Seurat input data, usually from Read10X().}

\item{datasetId}{This will be used as a prefix for barcodes, and stored in metadata. Also used as the project name for the Seurat object.}

\item{datasetName}{An optional print-friendly name that will be stored in metadata}

\item{minFeatures}{Include cells where at least this many features are detected.}

\item{minCells}{Include features detected in at least this many cells.}

\item{mitoGenesPattern}{The expression to use when identifying mitochondrial genes}

\item{annotateMitoFromReference}{If true, a list of mitochondrial genes, taken from (https://www.genedx.com/wp-content/uploads/crm_docs/Mito-Gene-List.pdf) will be used to calculate p.mito}
}
\value{
A Seurat object with p.mito calculated.
}
\description{
Create Seurat object from count data (usually from Read10X()). This also sets pct.mito.
}
