% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{CreateSeuratObj}
\alias{CreateSeuratObj}
\title{Wrapper around Seurat::CreateSeuratObject}
\usage{
CreateSeuratObj(
  seuratData,
  project,
  minFeatures = 25,
  minCells = 0,
  mitoGenesPattern = "^MT-",
  gtfFile = NULL
)
}
\arguments{
\item{seuratData, }{Seurat input data, usually from Read10X().}

\item{project, }{Sets the project name for the Seurat object.}

\item{minFeatures, }{Include cells where at least this many features are detected.}

\item{minCells, }{Include features detected in at least this many cells.}

\item{mitoGenesPattern}{The expression to use when identfying mitochondrial genes}

\item{gtfFile}{Ths optional GTF file used to create these data. If provided, it will be scanned for genes from the MT contig, and used to compute p.mito}
}
\value{
A Seurat object with p.mito calculated.
}
\description{
Create Seurat object from count data (usually from Read10X()). This also sets pct.mito.
}
