% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{DesignModelMatrix}
\alias{DesignModelMatrix}
\title{DesignModelMatrix}
\usage{
DesignModelMatrix(
  seuratObj,
  contrast_columns,
  sampleIdCol = "cDNA_ID",
  test.use = "QLF",
  assayName = "RNA",
  minCountsPerGene = 1
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{contrast_columns}{A vector of columns to contrast}

\item{sampleIdCol}{An additional column denoting the variable containing the sample (for grouping)}

\item{assayName}{The name of the assay to use}

\item{minCountsPerGene}{Any genes with fewer than this many counts (across samples) will be dropped}

\item{test.use.}{Can be either QLF or LRT. QLF runs edgeR::glmQLFTest, while LRT runs edgeR::glmLRT}
}
\value{
An edgeR glm object
}
\description{
Creates an edgeR glm object
}
