% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{DesignModelMatrix}
\alias{DesignModelMatrix}
\title{DesignModelMatrix}
\usage{
DesignModelMatrix(seuratObj, contrast_columns, sampleIdCol = "cDNA_ID")
}
\arguments{
\item{seuratObj}{The seurat object}

\item{contrast_columns}{A vector of columns to contrast}

\item{sampleIdCol}{An additional column denoting the variable containing the sample (for grouping)}
}
\value{
A model matrix (samples x groups data frame.)
}
\description{
This creates a model matrix that groups the samples according to the experimental groups used downstream in statistical analyses.
}
