% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{.addRegulationInformationAndFilterDEGs}
\alias{.addRegulationInformationAndFilterDEGs}
\title{.addRegulationInformationAndFilterDEGs}
\usage{
.addRegulationInformationAndFilterDEGs(
  tibble,
  logFC_threshold = 1,
  FDR_threshold = 0.05
)
}
\arguments{
\item{tibble}{a (presumably large) tibble from dplyr containing the results of RunFilteredPseudobulkingContrasts}
}
\value{
a tibble with the uniqueness and direction of regulation for each gene stored in tibble$regulation
}
\description{
a helper function for PseudobulkingBarPlot to tag genes as up- or down-regulated and check for uniqueness of differential gene expression across the contrasts.
}
