% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssGSEA_Alternatives.R
\name{.AlternativeSsgseaSeurat}
\alias{.AlternativeSsgseaSeurat}
\title{Alternative ssGSEA implementation for Seurat objects}
\usage{
.AlternativeSsgseaSeurat(
  seuratObj = seuratObj,
  geneSets,
  method = "dts",
  groupSize = 5000,
  assay = "RNA",
  layer = "counts",
  outputAssayName = "ssGSEA.alternative",
  parallelWorkers = 4,
  timing = FALSE
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{geneSets}{A list of gene sets to run ssGSEA on. The gene sets should be named lists of gene symbols.}

\item{method}{The method to use for ssGSEA. The dts, wass, and ks methods from the twosamples package are supported.}

\item{groupSize}{The number of groups to split the data into for parallel processing.}

\item{assay}{The assay to use.}

\item{layer}{The layer to use.}

\item{outputAssayName}{The name of the assay to store the results in.}

\item{parallelWorkers}{The number of parallel workers/cores to use.}

\item{timing}{If TRUE, the time taken to run the function will be printed.}
}
\value{
A Seurat object with the ssGSEA scores stored in the specified assay.
}
\description{
This function will run a modified version of ssGSEA.
}
\examples{
\dontrun{
# Load the Seurat object
data("pbmc_small")
seuratObj <- pbmc_small
# Get geneSets
geneSets <- escape::getGeneSets(library = "H")
# Run the alternative ssGSEA
seuratObj <- .AlternativeSsgseaSeurat(seuratObj, geneSets = geneSets)
}
}
