% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADTs.R
\name{.CalculateStatistics}
\alias{.CalculateStatistics}
\title{Calculate Statistics for ADT distributions}
\usage{
.CalculateStatistics(
  seuratObj,
  whitelist,
  adtwhitelist,
  assay,
  layer,
  libraryIdentificationColumn,
  minimumCounts,
  minimumCells,
  plots,
  verboseplots,
  peakdist_sd_ratio,
  peakheight_ratio_threshold,
  troughheight_ratio_threshold,
  xdist_ratio_threshold,
  antimode_loc_min,
  min_frac_under_threshold,
  max_frac_under_threshold
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{whitelist}{A vector of library IDs to run the classification on. If NULL, all libraries will be run.}

\item{adtwhitelist}{An optional vector of ADTs to be triaged and classified.}

\item{assay}{The assay to use.}

\item{layer}{The layer to use.}

\item{libraryIdentificationColumn}{The column in the metadata that identifies the atomic unit of sequencing data (e.g. cDNA library).}

\item{minimumCounts}{A lower bound on the number of reads in the ADT assay to attempt classification.}

\item{minimumCells}{A lower bound on the number of cells recovered from a library to attempt classification.}

\item{plots}{If TRUE, plots will be generated.}

\item{verboseplots}{if TRUE, intermediate plots will be generated}

\item{peakdist_sd_ratio}{An upper bound of acceptable peak distance divided by the standard deviation of the ADT.}

\item{peakheight_ratio_threshold}{The threshold for the peak height ratio.}

\item{troughheight_ratio_threshold}{The threshold for the peak2 - antimode to peak2 ratio.}

\item{xdist_ratio_threshold}{The threshold for the peak1_antimode_distance/peak2_antimode_distance ratio.}

\item{antimode_loc_min}{The minimum antimode location.}

\item{min_frac_under_threshold}{The minimum fraction of cells under the antimode.}

\item{max_frac_under_threshold}{The minimum fraction of cells above the antimode.}
}
\description{
This function will calculate metrics and produce a dataframe that will be used for classification by .TriageADTs
}
