% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADTs.R
\name{.TriageADTs}
\alias{.TriageADTs}
\title{Triage ADTs for classification}
\usage{
.TriageADTs(
  df,
  minimumCells,
  peakdist_sd_ratio,
  peakheight_ratio_threshold,
  xdist_ratio_threshold,
  troughheight_ratio_threshold,
  plots,
  antimode_loc_min,
  min_frac_under_threshold,
  max_frac_under_threshold
)
}
\arguments{
\item{df}{A dataframe resulting from running .CalculateStatistics on a Seurat object}

\item{minimumCells}{A lower bound on the number of cells recovered from a library to attempt classification.}

\item{peakdist_sd_ratio}{An upper bound of acceptable peak distance divided by the standard deviation of the ADT.}

\item{peakheight_ratio_threshold}{The threshold for the peak height ratio.}

\item{xdist_ratio_threshold}{The threshold for the peak1_antimode_distance/peak2_antimode_distance ratio.}

\item{troughheight_ratio_threshold}{The threshold for the peak2 - antimode to peak2 ratio.}

\item{plots}{If TRUE, plots will be generated.}

\item{antimode_loc_min}{The minimum antimode location.}

\item{min_frac_under_threshold}{The minimum fraction of cells under the antimode.}

\item{max_frac_under_threshold}{The minimum fraction of cells above the antimode.}
}
\description{
This function will take in a dataframe and make calls on bimodality
}
