% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{FilterPseudobulkContrasts}
\alias{FilterPseudobulkContrasts}
\title{FilterPseudobulkContrasts}
\usage{
FilterPseudobulkContrasts(
  logical_dataframe = NULL,
  design = NULL,
  contrast_columns = NULL,
  use_invariant_logic = T,
  invariant_fields = NULL,
  filtered_contrasts_output_file = "./filtered_contrasts.tsv"
)
}
\arguments{
\item{logical_dataframe}{The dataframe that defines the study design. Please see examples for the setup and use. This dataframe should have three columns: fields_to_check, field_logic, and criteria. fields_to_check defines the metadata field (e.g. Tissue or Timepoint) to which the logic gate will be applied. field_logic defines the logic gate that will be used for that field (one of: all (AND), any (OR), xor, nand, nor, or xnor. criteria defines the specific value of the metadata field that will be tested for equivalence against the positive and negative contrasts within the gate.}

\item{design}{a design/model matrix returned by DesignModelMatrix().}

\item{contrast_columns}{The vector of contrast names supplied to DesignModelMatrix() upstream of this function. Please ensure this vector is identical to the contrast_columns vector supplied to DesignModelMatrix().}

\item{use_invariant_logic}{Whether or not to apply invariance logic using invariant_fields. It is possible to use logic gates to keep metadata fields constant, but using invarant logic saves time & effort in those cases.}

\item{invariant_fields}{The metadata columns of the SeuratObj that you wish to keep constant. For example: defining "Tissue" in this vector would filter all constrasts comparing Liver to PBMC, as their tissue is not invariant.}

\item{filtered_contrasts_output_file}{The file to write the list of filtered contrasts to (one pair of samples per row).}
}
\value{
A matrix of pairwise contrasts (number of contrasts x 2).
}
\description{
This is designed to accept a study design defined by a series of logical gates applied to metadata fields, test them for equivalence to some criterion, and then filter pairwise contrasts accordingly.
}
\examples{
#set up design matrix
contrast_columns = c("Timepoint", "Vaccine", "Challenge", "Tissue", "cell_type", "SampleType")
design <- DesignModelMatrix(seuratObj, contrast_columns = c("Timepoint", "Vaccine", "Challenge", "Tissue", "cell_type","SampleType"), sampleIdCol = 'cDNA_ID')

#form study design using logic gates
fields_to_check <- c("Challenge", "SampleType")
field_logic <- c("xor", "all") #check for equality. any can be used for an OR gate, all can be used as an AND gate. 
criteria <- c("Mock.challenged", "Necropsy")
logical_dataframe <- data.frame(fields_to_check = fields_to_check, field_logic = field_logic, criteria = criteria)
invariant_fields <- c("Tissue","cell_type", "SampleType")

#finally, enumerate all possible contrasts, then filter and return them according to the study design defined by logical_dataframe.
filtered_contrasts <- FilterPseudobulkContrasts(logical_dataframe = logical_dataframe, design = design, contrast_columns = contrast_columns, use_invariant_logic = T, invariant_fields = invariant_fields, filtered_contrasts_output_file = "./filtered_contrasts.tsv")
}
