% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{FilterRawCounts}
\alias{FilterRawCounts}
\title{Perform Cell Filtering on the cells of the seurat object}
\usage{
FilterRawCounts(
  seuratObj,
  nCount_RNA.high = 20000,
  nCount_RNA.low = 1,
  nFeature.high = 3000,
  nFeature.low = 200,
  pMito.high = 0.15,
  pMito.low = 0
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{nCount_RNA.high}{Cells with nCount_RNA greater than this value will be filtered}

\item{nCount_RNA.low}{Cells with nCount_RNA less than this value will be filtered}

\item{nFeature.high}{Cells with nFeature greater than this value will be filtered}

\item{nFeature.low}{Cells with nFeature less than this value will be filtered}

\item{pMito.high}{Cells with percent mito greater than this value will be filtered}

\item{pMito.low}{Cells with percent mito  less than this value will be filtered}
}
\value{
The modified seurat object
}
\description{
This will perform filtering on the cells of the seurat object
}
