% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{Find_Markers}
\alias{Find_Markers}
\title{Find_Markers}
\usage{
Find_Markers(
  seuratObj,
  resolutionToUse,
  outFile = NULL,
  testsToUse = c("wilcox", "bimod", "roc", "t", "negbinom", "poisson", "LR", "MAST",
    "DESeq2"),
  numGenesToSave = 20,
  onlyPos = F
)
}
\arguments{
\item{seuratObj}{A seurat object}

\item{resolutionToUse}{The resolution, generally computed during FindClustersAndDimRedux()}

\item{outFile}{A file where a table of markers will be saved}

\item{testsToUse}{A vector of tests to be used.  Each will be used to run FindAllMarkers() and the results merged}

\item{numGenesToSave}{The number of top markers per cluster to save}

\item{onlyPos}{If true, only positive markers will be saved}
}
\value{
A DT::datatable object with the top markers, suitable for printing
}
\description{
Find_Markers
}
