% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{FindClustersAndDimRedux}
\alias{FindClustersAndDimRedux}
\title{Find Clusters And Perform DimRedux}
\usage{
FindClustersAndDimRedux(
  seuratObj,
  dimsToUse = NULL,
  minDimsToUse = NULL,
  umap.method = "uwot",
  umap.n.neighbors = 30L,
  umap.min.dist = 0.3,
  umap.seed = 1234,
  umap.n.epochs = NULL,
  max.tsne.iter = 10000,
  tsne.perplexity = 30,
  clusterResolutions = c(0.2, 0.4, 0.6, 0.8, 1.2)
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{dimsToUse}{The number of dims to use.  If null, this will be inferred using FindSeuratElbow()}

\item{minDimsToUse}{The minimum numbers of dims to use.  If dimsToUse is provided, this will override.}

\item{umap.method}{The UMAP method, either uwot or umap-learn, passed directly to Seurat::RunUMAP}

\item{umap.n.neighbors}{Passed directly to Seurat::RunUMAP}

\item{umap.min.dist}{Passed directly to Seurat::RunUMAP}

\item{umap.seed}{Passed directly to Seurat::RunUMAP}

\item{umap.n.epochs}{Passed directly to Seurat::RunUMAP}

\item{max.tsne.iter}{The value of max_iter to provide to RunTSNE.  Increasing can help large datasets.}

\item{tsne.perplexity}{tSNE perplexity. Passed directly to Seurat::RunTSNE(), but CellMembrane:::.InferPerplexityFromSeuratObj() corrects it if need be based dataset dims.}

\item{clusterResolutions}{A vector of clustering resolutions, default is (0.2, 0.4, 0.6, 0.8, 1.2).}
}
\value{
A modified Seurat object.
}
\description{
Find Clusters And Perform DimRedux
}
