% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doublets.R
\name{FindDoublets}
\alias{FindDoublets}
\title{Find Doublets using scDblFinder}
\usage{
FindDoublets(seuratObj, assay = "RNA", rawResultFile = NULL, doPlot = TRUE)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{assay}{The assay to use}

\item{rawResultFile}{An optional path to a file where raw results will be written}

\item{doPlot}{If true, a DimPlot will be generated, assuming reductions are calculated}
}
\description{
Find Doublets using scDblFinder
}
