% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{HasSplitLayers}
\alias{HasSplitLayers}
\title{HasSplitLayers}
\usage{
HasSplitLayers(seuratObj, assaysToTest = NULL)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{assaysToTest}{An option list of assays to test}
}
\value{
A boolean indicating whether the object has split layers
}
\description{
HasSplitLayers
}
