% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{MergeSeuratObjs}
\alias{MergeSeuratObjs}
\title{Merge Seurat Objects}
\usage{
MergeSeuratObjs(
  seuratObjs,
  projectName,
  assay = NULL,
  normalization.method = "LogNormalize"
)
}
\arguments{
\item{seuratObjs}{A named list of seurat objects, optionally named (in which case these will be used as dataset names).}

\item{projectName}{The project name when creating the final seuratObj}

\item{assay}{The assay to use}

\item{normalization.method}{Normalization method}
}
\value{
A modified Seurat object.
}
\description{
Merges a list of Seurat objects
}
