% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{MergeSeuratObjs}
\alias{MergeSeuratObjs}
\title{Merge Seurat Objects}
\usage{
MergeSeuratObjs(
  seuratObjs,
  projectName,
  merge.data = FALSE,
  expectedDefaultAssay = "RNA",
  enforceUniqueCells = TRUE,
  errorOnBarcodeSuffix = FALSE,
  doGC = FALSE
)
}
\arguments{
\item{seuratObjs}{A named list of seurat objects, optionally named (in which case these will be used as dataset names).}

\item{projectName}{The project name when creating the final seurat object}

\item{merge.data}{Passed directly to Seurat::merge}

\item{expectedDefaultAssay}{If not null, the DefaultAssay on the resulting seurat object will be set to this}

\item{enforceUniqueCells}{If true, all inputs must have unique cellbarcodes.}

\item{errorOnBarcodeSuffix}{In certain cases, software appends a digit (i.e. -1) to the end of cellbarcodes. These can be a problem when trying to make string comparisons. If true, the method will error if these are encountered.}

\item{doGC}{If true, in an attempt to save memory gc() will be run after each seurat object is merged}
}
\value{
A modified Seurat object.
}
\description{
Merges a list of Seurat objects
}
