% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{NormalizeAndScale}
\alias{NormalizeAndScale}
\title{Run the primary seurat processing steps.}
\usage{
NormalizeAndScale(
  seuratObj,
  nVariableFeatures = NULL,
  block.size = 1000,
  variableGenesWhitelist = NULL,
  variableGenesBlacklist = NULL,
  featuresToRegress = c(paste0("nCount_", Seurat::DefaultAssay(seuratObj))),
  scaleVariableFeaturesOnly = TRUE,
  includeCellCycleGenesInScaleData = TRUE,
  useSCTransform = FALSE,
  additionalFindVariableFeatureArgList = NULL,
  scoreCellCycle = TRUE,
  useAlternateG2M = FALSE
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{nVariableFeatures}{The number of variable features, passed to either FindVariableFeatures or SCTransform}

\item{block.size}{Passed directly to ScaleData}

\item{variableGenesWhitelist}{An optional vector of genes that will be included in PCA, beyond the default VariableFeatures()}

\item{variableGenesBlacklist}{An optional vector of genes that will be excluded from PCA, beyond the default VariableFeatures()}

\item{featuresToRegress}{The set of features which will be passed to Seurat::ScaleData vars.to.regress}

\item{scaleVariableFeaturesOnly}{If true, ScaleData will only be performed on VariableFeatures(), which is governed by FindVariableFeatures, variableGenesWhitelist, and variableGenesBlacklist}

\item{includeCellCycleGenesInScaleData}{If true, cell cycle genes will always be included in the features passed to ScaleData().}

\item{useSCTransform}{If true, SCTransform will be used in place of the standard Seurat workflow (NormalizeData, ScaleData, FindVariableFeatures)}

\item{additionalFindVariableFeatureArgList}{A list of arguments passed directly to FindVariableFeatures}

\item{scoreCellCycle}{If true, ScoreCellCycle will be run to compute Phase, which is stored in meta.data. If a field named Phase already exists, this will be skipped.}

\item{useAlternateG2M}{If true, this will use a smaller set of G2M genes, defined from: https://raw.githubusercontent.com/hbc/tinyatlas/master/cell_cycle/Homo_sapiens.csv}
}
\value{
A modified Seurat object.
}
\description{
This is the primary entry point for processing scRNAseq data with Seurat
}
