% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{NormalizeAndScale}
\alias{NormalizeAndScale}
\title{Run the primary seurat processing steps.}
\usage{
NormalizeAndScale(
  seuratObj,
  variableFeatureSelectionMethod = "vst",
  nVariableFeatures = 2000,
  mean.cutoff = c(0.0125, 3),
  dispersion.cutoff = c(0.5, Inf)
)
}
\arguments{
\item{seuratObj, }{A Seurat object.}

\item{variableFeatureSelectionMethod}{The selection method to be passed to FindVariableFeatures()}

\item{nVariableFeatures}{The number of variable features to find}

\item{mean.cutoff}{Passed directly to FindVariableFeatures}

\item{dispersion.cutoff}{Passed directly to FindVariableFeatures}
}
\value{
A modified Seurat object.
}
\description{
This is the primary entry point for processing scRNAseq data with Seurat
}
