% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{PerformEmptyDropletFiltering}
\alias{PerformEmptyDropletFiltering}
\title{PerformEmptyDropletFiltering}
\usage{
PerformEmptyDropletFiltering(
  seuratRawData,
  fdrThreshold = 0.01,
  emptyDropNIters = 10000,
  emptyDropsLower = 100
)
}
\arguments{
\item{seuratRawData}{Raw data}

\item{fdrThreshold}{FDR threshold, passed directly to PerformEmptyDrops()}

\item{emptyDropNIters}{Number of iterations, passed directly to PerformEmptyDrops()}

\item{emptyDropsLower}{Passed directly to emptyDrops(). The lower bound on the total UMI count, at or below which all barcodes are assumed to correspond to empty droplets.}
}
\value{
Plot
}
\description{
PerformEmptyDropletFiltering
}
