% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CiteSeq.R
\name{PlotAverageAdtCounts}
\alias{PlotAverageAdtCounts}
\title{Plot Average ADT counts}
\usage{
PlotAverageAdtCounts(
  seuratObj,
  groupFields = c("ClusterNames_0.2", "ClusterNames_0.4", "ClusterNames_0.6"),
  assayName = "ADT",
  layer = "counts",
  outFile = NA,
  normalization.method = "CLR",
  margin = 1
)
}
\arguments{
\item{seuratObj}{The seurat object where data will be added.}

\item{groupFields}{The directory holding raw count data, generally the raw_feature_bc_matrix from the cellranger outs folder}

\item{assayName}{The name of the assay holding data}

\item{layer}{The assay layer to use for average expression data}

\item{outFile}{If provided, the heatmap will be written to this file}

\item{normalization.method}{The normalization method to apply after AverageExpression(). If null, this will be skipped.}

\item{margin}{This is provided to Seurat::NormalizeData()}
}
\description{
Creates a heatmap with average counts of each ADT markers based on one or more grouping fields
}
