% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phenotyping.R
\name{PlotMarkerSeries}
\alias{PlotMarkerSeries}
\title{PlotMarkerSeries}
\usage{
PlotMarkerSeries(
  seuratObj,
  features,
  reductions = c("umap"),
  title = NULL,
  setSize = 4
)
}
\arguments{
\item{features}{A vector of feature names}

\item{reductions}{The reductions to plot}

\item{title}{An optional title of this plot series}

\item{setSize}{The maximum number of features to include per FeaturePlot}
}
\description{
Iteratively plots a set of markers
}
