% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{PlotSeuratVariables}
\alias{PlotSeuratVariables}
\title{Plot Seurat Variables}
\usage{
PlotSeuratVariables(seuratObj, xvar, yvar)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{xvar}{The first variable to summarize (i.e. ClusterNames_0.2)}

\item{yvar}{The second variable to summarize (i.e. SubjectId)}
}
\description{
Create a panel of plots summarizing two variables within the seurat object
}
