% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctour.R
\name{PredictScTourPseudotime}
\alias{PredictScTourPseudotime}
\title{PredictScTourPseudotime}
\usage{
PredictScTourPseudotime(
  seuratObj = NULL,
  GEXOutfile = NULL,
  modelFile = NULL,
  variableGenesFile = NULL,
  ptimeOutFile = NULL,
  embeddingOutFile = NULL,
  assayName = "RNA",
  cleanUpIntermediateFiles = T
)
}
\arguments{
\item{seuratObj}{The Seurat object containing the data.}

\item{GEXOutfile}{The GEX filename used for output of DropletUtils::write10xCounts (.h5 extension).}

\item{modelFile}{A path pointing to a trained scTour model.}

\item{variableGenesFile}{A path pointing to a trained scTour model's variable features. An scTour model requires both the trained model and feature space of the training data. This file stores the training data's feature space as a (N+1)x1 csv where N is the number of variable genes after exclusion list subtraction and the first row notes the column name "gene_ids".}

\item{ptimeOutFile}{An output file that will contain a Nx2 csv where n is the number of cells in the seurat object, the first column is cell barcodes, and the second column is pseudotime as predicted by the model.}

\item{embeddingOutFile}{The scTour model will also compute cell embeddings and append them to the seurat object. These will be appended under the 'SCTOUR_" dimensional reduction.}

\item{assayName}{Assay whose data is to be written out with DropletUtils::write10xCounts. Should be "RNA".}

\item{cleanUpIntermediateFiles}{This boolean controls if GEXOutfile, embeddingOutFile, and ptimeOutFile should be deleted after model training.}
}
\value{
A Seurat Object with pseudotime from the supplied model written into its metadata.
}
\description{
Predicts pseudotime from a trained sctour model.
}
\examples{
\dontrun{

seuratObj <- PredictScTourPseudotime(seuratObj,
                                     GEXOutfile = './gex_tempfile.h5',
                                     modelFile = "./test_model.pth",
                                     variableGenesFile = './variable_genes_out_file.csv',
                                     embeddingOutFile = "./embeddings.csv",
                                     ptimeOutFile ='./ptime_out_file.csv', 
                                     cleanUpIntermediateFiles = T)
}
}
