% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{PseudobulkingBarPlot}
\alias{PseudobulkingBarPlot}
\title{PseudobulkingBarPlot}
\usage{
PseudobulkingBarPlot(
  filteredContrastsResults,
  metadataFilterList = NULL,
  title = "Please Title The Bar Plot",
  log_y_axis = FALSE,
  logFC_threshold = 1,
  FDR_threshold = 0.05,
  swapContrastDirectionality = FALSE
)
}
\arguments{
\item{filteredContrastsResults}{A list of dataframes returned by RunFilteredContrasts.}

\item{metadataFilterList}{An optional list of lists specifying further filtering to be performed. These lists must follow the format: list( list("filterDirection", "filterField", "filterValue" )) where: filterDirection determines whether the Positive, Negative, or Both sides of the contrasts should be filtered, filterField determines which metadata field (originally supplied to groupFields in PseudobulkSeurat and contrast_columns in DesignModelMatrix), filterValue corresponds to which values of the filterField should be filtered. This is useful if you passed parallel hypotheses (e.g. what genes are differentially expressed in each tissue?) in the requireIdenticalFields argument of RunFilteredContrasts, but want to plot the results of only one tissue at a time.}

\item{title}{Title for the bar plot.}

\item{log_y_axis}{A boolean determining if the y axis (magnitude of differential expression) should be log transformed.}

\item{logFC_threshold}{A passthrough argument specifying the log fold change threshold to be used by .addRegulationInformationAndFilterDEGs to filter genes and determine regulation direction.}

\item{FDR_threshold}{A passthrough argument specifying the FDR threshold to be used by .addRegulationInformationAndFilterDEGs to filter genes and determine regulation direction.}

\item{swapContrastDirectionality}{A boolean determining if the contrast directionality should be swapped. This is useful if you want the "control" condition in your contrasts to appear in the opposite directionality of the default.}
}
\value{
A list containing the filtered dataframe used for plotting and the bar plot itself.
}
\description{
This is a plotting function downstream of RunFilteredContrasts that will create a  bar plot with the results of each contrast ordered by the magnitude of differentially expressed genes.
}
