% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{PseudobulkSeurat}
\alias{PseudobulkSeurat}
\title{Pseudobulk Seurat}
\usage{
PseudobulkSeurat(seuratObj, groupFields, assays = NULL)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{groupFields}{The set of fields on which to group}

\item{assays}{The assays to aggregate}
}
\value{
An aggregated Seurat object.
}
\description{
Aggregates raw counts in the seurat object, generating a new seurat object where each same has the sum of the counts, grouped by the desired variables
}
