% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoBulk.R
\name{PseudobulkSeurat}
\alias{PseudobulkSeurat}
\title{Pseudobulk Seurat}
\usage{
PseudobulkSeurat(
  seuratObj,
  groupFields,
  assays = NULL,
  additionalFieldsToAggregate = NULL,
  metaFieldCollapseCharacter = "|",
  nCountRnaStratification = F,
  stratificationGroupingFields = c("ClusterNames_0.2", "ClusterNames_0.4",
    "ClusterNames_0.6", "ClusterNames_0.8", "ClusterNames1.2")
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{groupFields}{The set of fields on which to group}

\item{assays}{The assays to aggregate}

\item{additionalFieldsToAggregate}{An option list of additional fields (which must be numeric). Per field, the mean per group will be computed and stored in the output object.}

\item{metaFieldCollapseCharacter}{The character to use when concatenating metadata fields together to form the sample key field}

\item{nCountRnaStratification}{A boolean determining whether or not automatic outlier detection of clusters with abnormal nCount_RNA should be detected.}

\item{stratificationGroupingFields}{The metadata fields containing groupings of cells to be specified as having abnormal nCount_RNA distributions or not. The "leftmost"/first argument that detects an outlier will break the looping and sequester cells at that value. It is recommended that this vector is arranged by a "least to most" specific hierarchy/progression.}
}
\value{
An aggregated Seurat object.
}
\description{
Aggregates raw counts in the seurat object, generating a new seurat object where each same has the sum of the counts, grouped by the desired variables
}
