% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{ReadAndFilter10xData}
\alias{ReadAndFilter10xData}
\title{Read and Filter 10X files.}
\usage{
ReadAndFilter10xData(
  dataDir,
  datasetName,
  emptyDropNIters = 10000,
  storeGeneIds = TRUE,
  emptyDropsLower = 100
)
}
\arguments{
\item{dataDir}{The directory holding raw count data}

\item{datasetName}{A name to use when creating the Seurat object}

\item{emptyDropNIters}{The number of iterations to use with PerformEmptyDrops()}

\item{storeGeneIds}{If true, a map to translate geneId and name (by default rownames will use gene name)}

\item{emptyDropsLower}{Passed directly to emptyDrops(). The lower bound on the total UMI count, at or below which all barcodes are assumed to correspond to empty droplets.}
}
\value{
A Seurat object.
}
\description{
Reads in 10X files using Read10X and filters abberent cells using PerformEmptyDropletFiltering and returns a Seurat object.
}
