% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{RemoveCellCycle}
\alias{RemoveCellCycle}
\title{Remove Cell Cycle}
\usage{
RemoveCellCycle(seuratObj, min.genes = 10, block.size = 1000)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{min.genes}{If less than min.genes are shared between the seurat object and the reference cell cycle genes, this method will abort.}

\item{block.size}{Passed directly to ScaleData}

\item{scaleVariableFeaturesOnly}{If true, ScaleData will only be performed on genes specified in FindVariableFeatures()}
}
\value{
A modified Seurat object.
}
\description{
Remove Cell Cycle
}
