% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{RunCoNGA}
\alias{RunCoNGA}
\title{Run CoNGA}
\usage{
RunCoNGA(
  seuratObj = NULL,
  tcrFileFromRdiscvr = NULL,
  seuratToCongaDir = "./seuratToConga",
  assayName = "RNA",
  organism = NULL,
  runCongaOutputFilePrefix = "conga_output",
  gexDatatype = "10x_h5",
  runCongaOutfilePrefixForQcPlots = "qc_plots",
  runCongaOutputDirectory = "./conga_output",
  congaMetadataPrefix = "conga_"
)
}
\arguments{
\item{seuratObj}{The Seurat object containing the data to be run using CoNGA.}

\item{tcrFileFromRdiscvr}{The 10x clonotypes file for this Seurat Object created by Rdiscvr::CreateMergedTcrClonotypeFile().}

\item{seuratToCongaDir}{The directory to store the results of .SeuratToConga() (the input files for the python call to run_CoNGA()).}

\item{assayName}{Pass-through variable for accessing the assay name within the Seurat object for .SeuratToConga().}

\item{organism}{'human' or 'rhesus'}

\item{runCongaOutputFilePrefix}{prefix for the output files from the python call to run_CoNGA().}

\item{gexDatatype}{This should be "10x_h5" since we're using DropletUtils to write the counts, although "h5ad" is supported.}

\item{runCongaOutfilePrefixForQcPlots}{Prefix for the qc output files to be generated by run_CoNGA().}

\item{runCongaOutputDirectory}{The directory that will store the many files created during run_CoNGA().}

\item{congaMetadataPrefix}{A prefix to be added to the columns of the metadata that will be added from CoNGA within the returned Seurat object.}
}
\value{
A Seurat object with the conga metadata relevant to TCR + gene expression clustering appended.
}
\description{
Runs CoNGA on a seurat object
}
\examples{
\dontrun{
  CreateMergedTcrClonotypeFile(seuratObj,
                               outputFile = "./tcrs.csv")
  congaSeuratObj <- RunCoNGA(seuratObj = seuratObj,
                             tcrFileFromRdiscvr = "./tcrs.csv",
                             organism = "rhesus")
}
}
