% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{RunCoNGA}
\alias{RunCoNGA}
\title{Run CoNGA}
\usage{
RunCoNGA(
  variable_features_file,
  tcr_datafile,
  gex_datafile,
  organism,
  outfile_prefix,
  gex_datatype = "10x_h5",
  clones_file,
  outfile_prefix_for_qc_plots,
  working_directory = getwd()
)
}
\arguments{
\item{variable_features_file}{textfile with list of most variable genes. This should be outputDir(from SeuratToConga())/varfeats.csv.}

\item{tcr_datafile}{Merged TCR clonotype file created by SeuratToConga(). This should be outputDir(from SeuratToConga())/TCRs.csv.}

\item{gex_datafile}{h5 file created by SeuratToConga(). This should be outputDir(from SeuratToConga())/GEX.h5.}

\item{organism}{'human' or 'rhesus'}

\item{outfile_prefix}{prefix for the output files.}

\item{gex_datatype}{This should be "10x_h5" since we're using DropletUtils to write the counts, although "h5ad" is supported.}

\item{clones_file}{Clones file to be generated by RunCoNGA. This should a text file (e.g. "outputDir/clones_file.txt").}

\item{outfile_prefix_for_qc_plots}{Prefix for the qc output files.}

\item{working_directory}{The current working directory (to be passed into python to run CoNGA).}
}
\description{
Runs CoNGA on data generated by CellMembrane::SeuratToCoNGA()
}
