% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{RunCoNGA}
\alias{RunCoNGA}
\title{Run CoNGA}
\usage{
RunCoNGA(
  seuratObj = NULL,
  tcrClonesFile = NULL,
  seuratToCongaDir = "./seuratToConga",
  assayName = "RNA",
  organism = NULL,
  runCongaOutputFilePrefix = "conga_output",
  gexDatatype = "10x_h5",
  runCongaOutfilePrefixForQcPlots = "qc_plots",
  runCongaOutputDirectory = "./conga_output",
  congaMetadataPrefix = "conga_",
  pngConversionTool = NULL
)
}
\arguments{
\item{seuratObj}{The Seurat object containing the data to be run using CoNGA.}

\item{tcrClonesFile}{The 10x clonotypes file for this Seurat object. Single lanes can use the CellRanger/Vloupe contigs CSV file. Merged lanes need to merge these files and modify them to create unique cellbarcodes and clonotype names, such as the code from Rdiscvr::CreateMergedTcrClonotypeFile().}

\item{seuratToCongaDir}{The directory to store the results of SeuratToCoNGA() (the input files for the python call to run_CoNGA()).}

\item{assayName}{Pass-through variable for accessing the assay name within the Seurat object for SeuratToCoNGA().}

\item{organism}{'human' or 'rhesus'}

\item{runCongaOutputFilePrefix}{prefix for the output files from the python call to run_CoNGA().}

\item{gexDatatype}{This should be "10x_h5" since we're using DropletUtils to write the counts, although "h5ad" is supported.}

\item{runCongaOutfilePrefixForQcPlots}{Prefix for the qc output files to be generated by run_CoNGA().}

\item{runCongaOutputDirectory}{The directory that will store the many files created during run_CoNGA().}

\item{congaMetadataPrefix}{A prefix to be added to the columns of the metadata that will be added from CoNGA within the returned Seurat object.}

\item{pngConversionTool}{specify the conversion tool used by CoNGA for svg to png conversion. Can be any of: convert, inkscape, rsvg}
}
\value{
A Seurat object with the conga metadata relevant to TCR + gene expression clustering appended.
}
\description{
Runs CoNGA on a seurat object
}
\examples{
\dontrun{
  CreateMergedTcrClonotypeFile(seuratObj,
                               outputFile = "./tcrs.csv")
  congaSeuratObj <- RunCoNGA(seuratObj = seuratObj,
                             tcrClonesFile = "./tcrs.csv",
                             organism = "rhesus")
}
}
