% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape.R
\name{RunEscape}
\alias{RunEscape}
\title{Run escape ssGSEA}
\usage{
RunEscape(seuratObj, outputAssayName = "escape.ssGSEA", doPlot = FALSE)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{outputAssayName}{The name of the assay to store results}

\item{doPlot}{If true, a FeaturePlot will be printed for each pathway}
}
\value{
The seurat object with results stored in an assay
}
\description{
This will run escape to calcualte ssGSEA on Hallmark gene sets
}
