% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape.R
\name{RunEscape}
\alias{RunEscape}
\title{Run escape ssGSEA}
\usage{
RunEscape(
  seuratObj,
  outputAssayBaseName = "escape.",
  doPlot = FALSE,
  performDimRedux = FALSE,
  msigdbGeneSets = c("H", C5 = "GO:BP", C5 = "GO:MF", C5 = "GO:CC"),
  customGeneSets = NULL,
  customGeneSetAssayName = "CustomGeneSet",
  assayName = "RNA",
  maxBatchSize = 1e+05
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{outputAssayBaseName}{The basename of the assay to store results. Escape will be run once for each gene set.}

\item{doPlot}{If true, a FeaturePlot will be printed for each pathway}

\item{performDimRedux}{If true, the standard seurat PCA/FindClusters/UMAP process will be run on the escape data. This may be most useful when using a customGeneSet or a smaller set of features/pathways}

\item{msigdbGeneSets}{A vector containing gene set codes specifying which gene sets should be fetched from MSigDB and calculated. Some recommendations in increasing computation time: H (hallmark, 50 gene sets), C8 (scRNASeq cell type markers, 830 gene sets), C2 (curated pathways, 6366 gene sets), GO:BP (GO biological processes, 7658). Each item will be processed into a separate assay.}

\item{customGeneSets}{A (preferably named) list containing gene sets to be scored by escape.}

\item{customGeneSetAssayName}{The name for the output assay (prefixed with outputAssayBaseName) for any customGeneSets}

\item{assayName}{The name of the source assay}

\item{maxBatchSize}{If more than this many cells are in the object, it will be split into batches of this size and run in serial.}
}
\value{
The seurat object with results stored in an assay
}
\description{
This will run escape to calcualte ssGSEA on Hallmark gene sets
}
