% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_III.R
\name{RunPcaSteps}
\alias{RunPcaSteps}
\title{Run Seurat PCA}
\usage{
RunPcaSteps(
  seuratObj,
  variableGenesWhitelist = NULL,
  variableGenesBlacklist = NULL,
  npcs = 50,
  variableGeneTable = NULL
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{variableGenesWhitelist}{An optional vector of genes that will be included in PCA, beyond the default VariableFeatures()}

\item{variableGenesBlacklist}{An optional vector of genes that will be excluded from PCA, beyond the default VariableFeatures()}

\item{npcs}{Number of PCs to use for RunPCA()}

\item{variableGeneTable}{If provided, a table of variable genes will be written to this file}
}
\value{
A modified Seurat object.
}
\description{
Run Seurat PCA
}
