% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{SeuratToCoNGA}
\alias{SeuratToCoNGA}
\title{SeuratToCoNGA}
\usage{
SeuratToCoNGA(
  seuratObj,
  tcr_file_from_rdiscvr,
  outputDir,
  assayName = "RNA",
  overwrite = TRUE
)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{tcr_file_from_rdiscvr}{The 10x clonotypes file for this run created by Rdiscvr::CreateMergedTcrClonotypeFile().}

\item{outputDir}{The local path to write the output files. The critical files are}

\item{assayName}{The name of the gene expression data assay.}

\item{overwrite}{Boolean describing whether or not to overwrite the GEX info through write10xCounts().}
}
\description{
A wrapper function to prepare a Seurat object for Conga.
}
