% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoNGA.R
\name{SeuratToCoNGA}
\alias{SeuratToCoNGA}
\title{SeuratToCoNGA}
\usage{
SeuratToCoNGA(seuratObj, clonotypesFile, outputDir, assayName = "RNA")
}
\arguments{
\item{seuratObj}{The seurat object}

\item{clonotypesFile}{The 10x clonotypes file for this run}

\item{outputDir}{The local path to write the output files}

\item{assayName}{The name of the gene expression data assay}
}
\description{
A wrapper function to prepare a Seurat object for Conga.
}
