% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctour.R
\name{TrainSctourModel}
\alias{TrainSctourModel}
\title{TrainSctourModel}
\usage{
TrainSctourModel(
  seuratObj = NULL,
  GEXOutfile = NULL,
  exclusionList = NULL,
  exclusionJsonPath = NULL,
  modelBasePath = NULL,
  modelFileName = NULL,
  ptimeOutFile = NULL,
  variableGenesFile = NULL,
  embeddingOutFile = NULL,
  assayName = "RNA",
  cleanUpIntermediateFiles = T
)
}
\arguments{
\item{seuratObj}{The Seurat object containing the data.}

\item{GEXOutfile}{The GEX filename used for output of DropletUtils::write10xCounts (.h5 extension).}

\item{exclusionList}{List of genes to be excluded from variable genes in sctour. Usually a RIRA GeneSet.}

\item{exclusionJsonPath}{Filename for the file containing the gene exclusion list (.json extension)}

\item{modelBasePath}{A directory that will store the resulting pytorch model after model training.}

\item{modelFileName}{The model's file name. The full name of the model will be modelFileName.pth}

\item{variableGenesFile}{An scTour model requires both the trained model and feature space of the training data. This file stores the training data's feature space as a (N+1)x1 csv where N is the number of variable genes after exclusion list subtraction and the first row notes the column name "gene_ids".}

\item{embeddingOutFile}{The scTour model yields a dimensionally reduced space that a UMAP can be computed on. This file stores cell embeddings into this latent space, which are added as an "SCTOUR_" reduction into the input seurat object.}

\item{assayName}{Assay whose data is to be written out with DropletUtils::write10xCounts. Should be "RNA".}

\item{cleanUpIntermediateFiles}{This boolean controls if GEXOutfile, embeddingOutFile, and exclusionJsonPath should be deleted after model training.}
}
\value{
A seurat object with pseudotime and dimensional reductions computed by scTour
}
\description{
Trains an sctour model on a Seurat object
}
\examples{
\dontrun{
seuratObj <- TrainSctourModel(seuratObj = seuratObj, 
                GEXOutfile = './gex_tempfile.h5', 
                modelBasePath =  './', 
                modelFileName = "test_model",
                exclusionJsonPath = './exclusion_tempfile.json',
                ptimeOutFile = './ptime_out_file.csv',
                variableGenesFile = './variable_genes_out_file.csv',
                assayName = "RNA",
                embeddingOutFile = "./embeddings.csv",
                cleanUpIntermediateFiles = T)

}
}
