% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADTs.R
\name{triageADTsAndClassifyCells}
\alias{triageADTsAndClassifyCells}
\title{Triage ADTs and Classify Cells}
\usage{
triageADTsAndClassifyCells(
  seuratObj,
  adtToInspect = "CD4",
  assay = "ADT",
  layer = "data",
  libraryIdentificationColumn = "cDNA_ID",
  minimumCounts = 200,
  minimumCells = 20,
  plots = T,
  whitelist = NULL
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{adtToInspect}{The name of the ADT to inspect.}

\item{assay}{The assay to use.}

\item{layer}{The layer to use.}

\item{libraryIdentificationColumn}{The column in the metadata that identifies the atomic unit of sequencing data (e.g. cDNA library).}

\item{minimumCounts}{A lower bound on the number of reads in the ADT assay to attempt classification.}

\item{minimumCells}{A lower bound on the number of cells recovered from a library to attempt classification.}

\item{plots}{If TRUE, plots will be generated.}

\item{whitelist}{A vector of library IDs to run the classification on. If NULL, all libraries will be run.}
}
\description{
This function will classify cells within a Seurat object based on characteristics relating to the bimodality of an ADT.
}
