% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratUtils.R
\name{WriteCellBarcodes}
\alias{WriteCellBarcodes}
\title{Write Cell Barcodes}
\usage{
WriteCellBarcodes(seuratObj, file)
}
\arguments{
\item{seuratObj}{The seurat object}

\item{file}{The file to which barcodes will be written}
}
\description{
Writes a table of cell barcodes to the provided file
}
